Separate the represented pet and battle-pet responsibility (#607)
Sixth delivery of the refactor push. 111 methods move into 7 bounded
submodules under session/pets: pet, summoning, persistence, battle_pet,
battle_pet_slots, battle_pet_journal and battle_pet_publication. Every file is
under the 1,000-line review budget.

session/mod.rs drops from 47,905 to 45,001 production lines. Across the six
deliveries the root falls from 75,841 to 45,001 - 41 per cent retired into
navigable owners.

The classifier first produced two persistence modules of eight and twenty-two
lines. A one-method module is not a responsibility, so they were merged into a
single session/pets/persistence.rs covering both pet and battle-pet save and
load plans.

Methods that mention a pet but belong to spell, aura or item responsibilities -
all separated by #589, #597 and #601 - stay where they are.

Safety established by comparison: the impl surface is 1,453 methods before and
after, an identical multiset, and the ownership baseline reports 3,761
associated items unchanged. All 18 baseline differences are private to
pub(in crate::session); nothing became public. The three direct-registry
deltas cover the identical method set, re-attributed to their new files.

Verified: cargo test -p wow-world --lib 3,822 passed and 0 failed,
session-ownership-check --syntax-only PASS, check_architecture.py check and
self-test PASS, fmt and diff hygiene clean. Ceiling tightened.
