+++ crates/wow-world/src/session/mod.rs
@@ -18,6 +18,7 @@ pub use lifecycle::PlayerSaveOutcomeLikeCpp;
 mod effect_learning;
 pub mod mailbox;
 mod movement;
+mod pets;
 mod player_cast;
 mod player_items;
 mod quest;
@@ -11519,29 +11520,6 @@ impl WorldSession {
         canonical
     }
 
-    fn player_pet_guid_state_like_cpp(&self) -> Option<Option<ObjectGuid>> {
-        let canonical = self.with_owned_player_like_cpp(|player| player.gameplay_state().pet_guid);
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            return Some(self.represented_pet_guid_like_cpp);
-        }
-        canonical
-    }
-
-    fn set_player_pet_guid_like_cpp(&mut self, pet_guid: Option<ObjectGuid>) -> bool {
-        let canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player.gameplay_state_mut().pet_guid = pet_guid;
-            })
-            .is_some();
-        #[cfg(test)]
-        if canonical || self.player_handle_like_cpp.is_none() {
-            self.represented_pet_guid_like_cpp = pet_guid;
-            return true;
-        }
-        canonical
-    }
-
     fn add_canonical_attacker_like_cpp(&mut self, victim: ObjectGuid, attacker: ObjectGuid) {
         if self
             .mutate_canonical_player_by_guid_like_cpp(victim, |victim| {
@@ -14125,112 +14103,6 @@ impl WorldSession {
         self.persistence_ports_like_cpp.player.instance_lock.clone()
     }
 
-    /// Attach this session to the one canonical journal owner for its
-    /// Battle.net account. The attachment releases any held journal lease on
-    /// drop, matching C++ `WorldSession` teardown.
-    pub fn set_battle_pet_account_attachment_like_cpp(
-        &mut self,
-        attachment: BattlePetAccountAttachmentLikeCpp,
-    ) {
-        self.invalidate_canonical_player_spell_hit_aura_authority_like_cpp();
-        self.battle_pet_account_attachment_like_cpp = Some(attachment);
-    }
-
-    /// The #160 account owner and this session's journal lease id, when the
-    /// canonical journal is attached (issue #161 purchase saga).
-    pub(crate) fn battle_pet_account_owner_lease_like_cpp(
-        &self,
-    ) -> Option<(
-        Arc<crate::battle_pet_account::BattlePetAccountOwnerLikeCpp>,
-        crate::battle_pet_account::BattlePetLeaseIdLikeCpp,
-    )> {
-        self.battle_pet_account_attachment_like_cpp
-            .as_ref()
-            .map(|attachment| {
-                (
-                    Arc::clone(attachment.owner_like_cpp()),
-                    attachment.lease_id_like_cpp(),
-                )
-            })
-    }
-
-    /// Acquire this session's journal lease through the #160 attachment
-    /// (issue #161 admission/recovery).
-    pub(crate) async fn battle_pet_try_acquire_journal_lease_like_cpp(&self) -> bool {
-        let Some(attachment) = &self.battle_pet_account_attachment_like_cpp else {
-            return false;
-        };
-        attachment.try_acquire_lease_like_cpp().await
-    }
-
-    /// Set the world-DB battle-pet breed/quality selection store (#161).
-    #[cfg(test)]
-    pub fn set_battle_pet_selection_store_like_cpp(
-        &mut self,
-        store: Arc<wow_data::battle_pet_selection::BattlePetSelectionStoreLikeCpp>,
-    ) {
-        self.battle_pet_selection_store_like_cpp = Some(store);
-    }
-
-    #[cfg(test)]
-    pub(crate) fn battle_pet_selection_store_like_cpp(
-        &self,
-    ) -> Option<&Arc<wow_data::battle_pet_selection::BattlePetSelectionStoreLikeCpp>> {
-        self.battle_pet_selection_store_like_cpp.as_ref()
-    }
-
-    /// Install the SQLx-free Character durability port for the recoverable
-    /// #161 purchase saga. The composition root owns the MariaDB adapter.
-    pub fn set_battle_pet_purchase_persistence_port_like_cpp(
-        &mut self,
-        store: Arc<dyn wow_persistence::BattlePetPurchasePersistencePortLikeCpp>,
-    ) {
-        self.persistence_ports_like_cpp.player.battle_pet_purchase = Some(store);
-    }
-
-    pub(crate) fn battle_pet_purchase_store_like_cpp(
-        &self,
-    ) -> Option<Arc<dyn wow_persistence::BattlePetPurchasePersistencePortLikeCpp>> {
-        self.persistence_ports_like_cpp
-            .player
-            .battle_pet_purchase
-            .as_ref()
-            .map(Arc::clone)
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_battle_pet_purchase_selection_override_like_cpp(
-        &mut self,
-        selection: Option<wow_data::battle_pet_selection::BattlePetTrainerSelectionLikeCpp>,
-    ) {
-        self.battle_pet_purchase_selection_override_like_cpp = selection;
-    }
-
-    #[cfg(test)]
-    pub(crate) fn battle_pet_purchase_selection_override_like_cpp(
-        &self,
-    ) -> Option<wow_data::battle_pet_selection::BattlePetTrainerSelectionLikeCpp> {
-        self.battle_pet_purchase_selection_override_like_cpp
-    }
-
-    /// One cloned DB2 species row for admission-time materialization (#161).
-    pub(crate) fn battle_pet_species_entry_like_cpp(
-        &self,
-        species: u32,
-    ) -> Option<wow_data::BattlePetSpeciesEntry> {
-        if let Some(attachment) = &self.battle_pet_account_attachment_like_cpp {
-            return attachment.owner_like_cpp().species_entry_like_cpp(species);
-        }
-        #[cfg(test)]
-        return self
-            .battle_pet_species_store
-            .as_ref()
-            .and_then(|store| store.get(species))
-            .cloned();
-        #[cfg(not(test))]
-        None
-    }
-
     /// Quarantine this session after an unreconcilable battle-pet purchase
     /// COMMIT, mirroring the #159 money-persistence indeterminate boundary:
     /// normal payout admission stays closed and the client must relog.
@@ -14803,31 +14675,6 @@ impl WorldSession {
         self.toy_store.as_ref()
     }
 
-    #[cfg(test)]
-    pub fn set_battle_pet_breed_quality_store(&mut self, store: Arc<BattlePetBreedQualityStore>) {
-        self.battle_pet_breed_quality_store = Some(store);
-    }
-
-    #[cfg(test)]
-    pub fn set_battle_pet_breed_state_store(&mut self, store: Arc<BattlePetBreedStateStore>) {
-        self.battle_pet_breed_state_store = Some(store);
-    }
-
-    #[cfg(test)]
-    pub fn set_battle_pet_species_store(&mut self, store: Arc<BattlePetSpeciesStore>) {
-        self.battle_pet_species_store = Some(store);
-    }
-
-    #[cfg(test)]
-    pub fn set_battle_pet_species_state_store(&mut self, store: Arc<BattlePetSpeciesStateStore>) {
-        self.battle_pet_species_state_store = Some(store);
-    }
-
-    #[cfg(test)]
-    pub fn set_battle_pet_xp_game_table(&mut self, table: Arc<BattlePetXpGameTableLikeCpp>) {
-        self.battle_pet_xp_game_table = Some(table);
-    }
-
     pub fn set_combat_ratings_game_table(&mut self, table: Arc<CombatRatingsGameTableLikeCpp>) {
         self.combat_ratings_game_table = Some(table);
     }
@@ -14844,40 +14691,6 @@ impl WorldSession {
             .as_ref()
             .map(|table| table.rating_multiplier_like_cpp(u16::from(level), rating))
             .unwrap_or(1.0)
-    }
-
-    pub(crate) fn battle_pet_calculate_stats_like_cpp(
-        &self,
-        breed: u16,
-        species: u32,
-        quality: u8,
-        level: u16,
-    ) -> Option<RepresentedBattlePetCalculatedStatsLikeCpp> {
-        if let Some(attachment) = &self.battle_pet_account_attachment_like_cpp {
-            return attachment
-                .owner_like_cpp()
-                .calculate_stats_like_cpp(breed, species, quality, level);
-        }
-        #[cfg(test)]
-        {
-            let stats = calculate_battle_pet_stats_like_cpp(
-                breed,
-                species,
-                quality,
-                level,
-                self.battle_pet_breed_state_store.as_ref()?,
-                self.battle_pet_species_state_store.as_ref()?,
-                self.battle_pet_breed_quality_store.as_ref()?,
-            )?;
-
-            Some(RepresentedBattlePetCalculatedStatsLikeCpp {
-                max_health: stats.max_health,
-                power: stats.power,
-                speed: stats.speed,
-            })
-        }
-        #[cfg(not(test))]
-        None
     }
 
     pub(crate) fn player_collection_state_snapshot_like_cpp(
@@ -15818,42 +15631,6 @@ impl WorldSession {
             standing,
             self.friendship_rep_reaction_store.as_deref(),
         )
-    }
-
-    fn battle_pet_xp_per_level_like_cpp(&self, level: u16) -> Option<u16> {
-        let canonical = self
-            .battle_pet_account_attachment_like_cpp
-            .as_ref()
-            .and_then(|attachment| attachment.owner_like_cpp().xp_per_level_like_cpp(level));
-        #[cfg(test)]
-        return canonical.or_else(|| {
-            self.battle_pet_xp_game_table
-                .as_ref()
-                .and_then(|table| table.xp_per_level_like_cpp(level))
-                .or_else(|| {
-                    self.represented_battle_pet_xp_per_level_like_cpp
-                        .get(&level)
-                        .copied()
-                })
-        });
-        #[cfg(not(test))]
-        canonical
-    }
-
-    fn battle_pet_species_has_flag_like_cpp(&self, species: u32, flag: i32) -> bool {
-        if let Some(attachment) = &self.battle_pet_account_attachment_like_cpp {
-            return attachment
-                .owner_like_cpp()
-                .species_has_flag_like_cpp(species, flag);
-        }
-        #[cfg(test)]
-        return self
-            .battle_pet_species_store
-            .as_ref()
-            .and_then(|store| store.get(species))
-            .is_some_and(|entry| entry.has_flag_like_cpp(flag));
-        #[cfg(not(test))]
-        false
     }
 
     pub(crate) fn represented_faction_reaction_to_like_cpp(
@@ -23465,55 +23242,6 @@ impl WorldSession {
         });
     }
 
-    fn disable_pet_controls_on_mount_like_cpp(&mut self, react_state: u8, command_state: u8) {
-        let Some(pet_guid) = self.player_pet_guid_state_like_cpp().flatten() else {
-            return;
-        };
-
-        let Some((previous_react_state, _)) = self.canonical_pet_mode_state_like_cpp() else {
-            return;
-        };
-        if !self.update_player_pet_lifecycle_state_like_cpp(|state| {
-            state.temporary_mount_react_state = Some(previous_react_state);
-        }) {
-            return;
-        }
-        if !self.update_canonical_pet_mode_state_like_cpp(react_state, command_state) {
-            return;
-        }
-        self.send_packet(&wow_packet::packets::pet::PetMode {
-            pet_guid,
-            react_state,
-            command_state,
-            flag: 0,
-        });
-    }
-
-    fn enable_pet_controls_on_dismount_like_cpp(&mut self) {
-        if let Some(pet_guid) = self.player_pet_guid_state_like_cpp().flatten() {
-            if let Some((current_react_state, command_state)) =
-                self.canonical_pet_mode_state_like_cpp()
-            {
-                let react_state = self
-                    .player_pet_lifecycle_state_snapshot_like_cpp()
-                    .and_then(|state| state.temporary_mount_react_state)
-                    .unwrap_or(current_react_state);
-                if self.update_canonical_pet_mode_state_like_cpp(react_state, command_state) {
-                    self.send_packet(&wow_packet::packets::pet::PetMode {
-                        pet_guid,
-                        react_state,
-                        command_state,
-                        flag: 0,
-                    });
-                }
-            }
-        }
-
-        let _ = self.update_player_pet_lifecycle_state_like_cpp(|state| {
-            state.temporary_mount_react_state = None;
-        });
-    }
-
     fn update_player_collision_height_like_cpp(&mut self) {
         let Some((_, mount_display_id, object_scale)) =
             self.player_unit_presentation_snapshot_like_cpp()
@@ -23793,45 +23521,6 @@ impl WorldSession {
         let catalogs = self.area_trigger_catalogs_for_test_like_cpp();
         self.check_area_triggers_with_catalogs_like_cpp(&catalogs)
             .await;
-    }
-
-    fn unsummon_represented_pet_for_same_map_teleport_if_out_of_range_like_cpp(
-        &mut self,
-        destination: wow_core::Position,
-        options: TeleportToOptionsLikeCpp,
-    ) {
-        if options & TELE_TO_NOT_UNSUMMON_PET_LIKE_CPP != 0 {
-            return;
-        }
-        let Some(pet_guid) = self.player_pet_guid_state_like_cpp().flatten() else {
-            return;
-        };
-        let map_id = u32::from(self.player_map_id_like_cpp());
-        let instance_id = self
-            .current_canonical_player_map_key_like_cpp()
-            .map(|key| key.instance_id)
-            .unwrap_or(0);
-
-        let should_unsummon = {
-            let Some(manager) = self.canonical_map_manager.as_ref().cloned() else {
-                return;
-            };
-            let Ok(manager) = manager.lock() else {
-                return;
-            };
-            let Some(managed) = manager.find_map(map_id, instance_id) else {
-                return;
-            };
-            let visibility_range = managed.map().visibility_range();
-            managed
-                .map()
-                .get_pet(pet_guid)
-                .is_some_and(|pet| pet.distance_to_position(destination) > visibility_range)
-        };
-
-        if should_unsummon {
-            self.unsummon_represented_pet_temporary_if_any_like_cpp();
-        }
     }
 
     fn player_cannot_enter_target_map_like_cpp(&self, map_id: u32) -> Option<(u32, u8, i32)> {
@@ -25335,179 +25024,6 @@ impl WorldSession {
         }
     }
 
-    fn with_canonical_pet_like_cpp<R>(
-        &self,
-        pet_guid: ObjectGuid,
-        inspect: impl FnOnce(&Pet) -> R,
-    ) -> Option<R> {
-        let manager = self.canonical_map_manager.as_ref()?.lock().ok()?;
-        let mut inspect = Some(inspect);
-        let mut result = None;
-        manager.do_for_all_maps(|managed| {
-            if result.is_some() {
-                return;
-            }
-            if let Some(pet) = managed.map().get_typed_pet(pet_guid) {
-                result = Some(inspect.take().expect("pet inspector consumed once")(pet));
-            }
-        });
-        result
-    }
-
-    fn with_canonical_pet_mut_like_cpp<R>(
-        &self,
-        pet_guid: ObjectGuid,
-        mutate: impl FnOnce(&mut Pet) -> R,
-    ) -> Option<R> {
-        let mut manager = self.canonical_map_manager.as_ref()?.lock().ok()?;
-        let mut mutate = Some(mutate);
-        let mut result = None;
-        manager.do_for_all_maps_mut(|managed| {
-            if result.is_some() {
-                return;
-            }
-            if let Some(pet) = managed.map_mut().get_typed_pet_mut(pet_guid) {
-                result = Some(mutate.take().expect("pet mutation consumed once")(pet));
-            }
-        });
-        result
-    }
-
-    fn canonical_pet_mode_state_like_cpp(&self) -> Option<(u8, u8)> {
-        let pet_guid = self.player_pet_guid_state_like_cpp().flatten()?;
-        let canonical = self.with_canonical_pet_like_cpp(pet_guid, |pet| {
-            let react_state = pet.creature().react_state() as u8;
-            let command_state = pet
-                .creature()
-                .unit()
-                .subsystems()
-                .control
-                .charm_info
-                .as_ref()
-                .map_or(wow_packet::packets::pet::COMMAND_FOLLOW_LIKE_CPP, |info| {
-                    info.command_state
-                });
-            (react_state, command_state)
-        });
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            return Some((
-                self.represented_pet_react_state_like_cpp,
-                self.represented_pet_command_state_like_cpp,
-            ));
-        }
-        canonical
-    }
-
-    fn update_canonical_pet_mode_state_like_cpp(
-        &mut self,
-        react_state: u8,
-        command_state: u8,
-    ) -> bool {
-        let Some(pet_guid) = self.player_pet_guid_state_like_cpp().flatten() else {
-            return false;
-        };
-        let canonical = self
-            .with_canonical_pet_mut_like_cpp(pet_guid, |pet| {
-                pet.creature_mut()
-                    .set_react_state(react_state_from_db_like_cpp(react_state));
-                pet.creature_mut()
-                    .unit_mut()
-                    .subsystems_mut()
-                    .control
-                    .init_charm_info()
-                    .command_state = command_state;
-            })
-            .is_some();
-        #[cfg(test)]
-        if !canonical && self.player_handle_like_cpp.is_none() {
-            self.represented_pet_react_state_like_cpp = react_state;
-            self.represented_pet_command_state_like_cpp = command_state;
-            return true;
-        }
-        canonical
-    }
-
-    fn player_pet_lifecycle_state_snapshot_like_cpp(
-        &self,
-    ) -> Option<PlayerPetLifecycleStateLikeCpp> {
-        let canonical =
-            self.with_owned_player_like_cpp(|player| player.pet_lifecycle_state_like_cpp().clone());
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            return Some(PlayerPetLifecycleStateLikeCpp {
-                stable: self.represented_pet_stable_like_cpp.clone(),
-                character_rows_empty_authority_complete: self
-                    .represented_character_pet_rows_empty_authority_complete_like_cpp,
-                temporary_unsummoned_pet_number: self
-                    .represented_temporary_unsummoned_pet_number_like_cpp,
-                old_pet_spell: self.represented_old_pet_spell_like_cpp,
-                temporary_mount_react_state: self.temporary_mount_pet_react_state_like_cpp,
-            });
-        }
-        canonical
-    }
-
-    fn update_player_pet_lifecycle_state_like_cpp(
-        &mut self,
-        update: impl FnOnce(&mut PlayerPetLifecycleStateLikeCpp),
-    ) -> bool {
-        if self.player_handle_like_cpp.is_some() {
-            return self
-                .with_owned_player_mut_like_cpp(|player| {
-                    update(player.pet_lifecycle_state_mut_like_cpp())
-                })
-                .is_some();
-        }
-        #[cfg(test)]
-        {
-            let mut state = self
-                .player_pet_lifecycle_state_snapshot_like_cpp()
-                .unwrap_or_default();
-            update(&mut state);
-            self.represented_pet_stable_like_cpp = state.stable;
-            self.represented_character_pet_rows_empty_authority_complete_like_cpp =
-                state.character_rows_empty_authority_complete;
-            self.represented_temporary_unsummoned_pet_number_like_cpp =
-                state.temporary_unsummoned_pet_number;
-            self.represented_old_pet_spell_like_cpp = state.old_pet_spell;
-            self.temporary_mount_pet_react_state_like_cpp = state.temporary_mount_react_state;
-            true
-        }
-        #[cfg(not(test))]
-        {
-            let _ = update;
-            false
-        }
-    }
-
-    pub(crate) fn begin_represented_character_pet_authority_load_like_cpp(&mut self) {
-        self.pet_load_query_holder_rows_like_cpp.reset();
-        self.invalidate_represented_character_pet_empty_authority_like_cpp();
-    }
-
-    fn invalidate_represented_character_pet_empty_authority_like_cpp(&mut self) {
-        let _ = self.update_player_pet_lifecycle_state_like_cpp(|state| {
-            state.character_rows_empty_authority_complete = false;
-        });
-        self.invalidate_canonical_player_spell_hit_aura_authority_like_cpp();
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn set_represented_pet_mode_state_like_cpp(
-        &mut self,
-        pet_guid: Option<ObjectGuid>,
-        react_state: u8,
-        command_state: u8,
-    ) {
-        self.set_represented_pet_mode_state_with_spell_like_cpp(
-            pet_guid,
-            react_state,
-            command_state,
-            0,
-        );
-    }
-
     #[allow(dead_code)]
     pub(crate) fn set_represented_pet_mode_state_with_spell_like_cpp(
         &mut self,
@@ -25556,175 +25072,6 @@ impl WorldSession {
         let _ = self.update_player_pet_lifecycle_state_like_cpp(|state| {
             state.temporary_mount_react_state = None;
         });
-    }
-
-    pub(crate) fn set_represented_pet_stable_like_cpp(&mut self, stable: PetStable) {
-        self.invalidate_represented_character_pet_empty_authority_like_cpp();
-        let _ = self.update_player_pet_lifecycle_state_like_cpp(|state| state.stable = stable);
-    }
-
-    pub(crate) fn apply_represented_login_pet_talent_reset_like_cpp(&mut self) -> bool {
-        const AT_LOGIN_RESET_PET_TALENTS_LIKE_CPP: u16 = 0x010;
-
-        if !self
-            .resolved_represented_at_login_flags_like_cpp()
-            .is_some_and(|flags| (flags & AT_LOGIN_RESET_PET_TALENTS_LIKE_CPP) != 0)
-        {
-            return false;
-        }
-
-        self.invalidate_represented_character_pet_empty_authority_like_cpp();
-
-        if !self.update_player_pet_lifecycle_state_like_cpp(|state| {
-            for pet in state.stable.active_pets.iter_mut().flatten() {
-                pet.specialization_id = 0;
-            }
-            for pet in state.stable.stabled_pets.iter_mut().flatten() {
-                pet.specialization_id = 0;
-            }
-            for pet in &mut state.stable.unslotted_pets {
-                pet.specialization_id = 0;
-            }
-        }) {
-            return false;
-        }
-        self.pet_load_query_holder_rows_like_cpp.spells.clear();
-        true
-    }
-
-    /// C++ `Player::RemovePet(nullptr, PET_SAVE_NOT_IN_SLOT, true)`.
-    ///
-    /// Represented boundary: clears the active represented pet link, resets
-    /// `PetStable::CurrentPetIndex`, and removes the live typed pet from the
-    /// canonical map if present. Reagent return and exact `Pet::SavePetToDB`
-    /// remain owned by the future full pet/inventory persistence runtime.
-    pub(crate) fn remove_represented_pet_not_in_slot_like_cpp(&mut self) {
-        self.invalidate_represented_character_pet_empty_authority_like_cpp();
-        let pet_guid = self.player_pet_guid_state_like_cpp().flatten();
-        if let Some(pet_guid) = pet_guid
-            && let Some(manager) = self.canonical_map_manager.as_ref().map(Arc::clone)
-            && let Ok(mut manager) = manager.lock()
-        {
-            let mut removed = false;
-            manager.do_for_all_maps_mut(|managed| {
-                if removed {
-                    return;
-                }
-                match managed.map_mut().remove_from_map_like_cpp(pet_guid, false) {
-                    Ok(_) => removed = true,
-                    Err(wow_map::RemoveFromMapError::ObjectNotFound { .. }) => {}
-                    Err(_) => {}
-                }
-            });
-        }
-
-        if self.player_pet_guid_state_like_cpp().flatten().is_some() {
-            let _ = self.set_player_pet_guid_like_cpp(None);
-            #[cfg(test)]
-            {
-                self.represented_pet_created_by_spell_like_cpp = 0;
-                self.represented_pet_react_state_like_cpp =
-                    wow_packet::packets::pet::REACT_DEFENSIVE_LIKE_CPP;
-                self.represented_pet_command_state_like_cpp =
-                    wow_packet::packets::pet::COMMAND_FOLLOW_LIKE_CPP;
-            }
-            let _ = self.update_player_pet_lifecycle_state_like_cpp(|state| {
-                state.temporary_mount_react_state = None;
-            });
-            #[cfg(test)]
-            {
-                self.represented_pet_movement_speed_rates_like_cpp =
-                    [1.0; UnitMoveTypeLikeCpp::COUNT];
-            }
-        }
-        let _ = self.update_player_pet_lifecycle_state_like_cpp(|state| {
-            state.stable.current_pet_index = None;
-        });
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_pet_stable_current_index_like_cpp(&self) -> Option<u32> {
-        self.player_pet_lifecycle_state_snapshot_like_cpp()
-            .and_then(|state| state.stable.current_pet_index)
-    }
-
-    pub(crate) fn load_represented_pet_stable_rows_like_cpp(
-        &mut self,
-        summoned_pet_number: u32,
-        rows: impl IntoIterator<Item = CharacterPetStableRowLikeCpp>,
-    ) -> usize {
-        self.invalidate_represented_character_pet_empty_authority_like_cpp();
-        let mut stable = PetStable::default();
-        let mut loaded = 0usize;
-        let mut query_had_rows = false;
-
-        for row in rows {
-            query_had_rows = true;
-            let slot = row.slot;
-            let pet_info = PetStableInfo {
-                name: row.name,
-                action_bar: row.action_bar,
-                pet_number: row.pet_number,
-                creature_id: row.creature_id,
-                display_id: row.display_id,
-                experience: row.experience,
-                health: row.health,
-                mana: row.mana,
-                last_save_time: row.last_save_time,
-                created_by_spell_id: row.created_by_spell_id,
-                specialization_id: row.specialization_id,
-                level: row.level,
-                react_state: react_state_from_db_like_cpp(row.react_state),
-                pet_type: pet_type_from_db_like_cpp(row.pet_type),
-                was_renamed: row.was_renamed,
-            };
-
-            if PetSaveMode::is_active_slot(slot) {
-                let index = slot as usize;
-                if stable.active_pets.len() <= index {
-                    stable.active_pets.resize_with(index + 1, || None);
-                }
-                stable.active_pets[index] = Some(pet_info);
-            } else if PetSaveMode::is_stabled_slot(slot) {
-                let index = (slot - PetSaveMode::stable_slot(0)) as usize;
-                if stable.stabled_pets.len() <= index {
-                    stable.stabled_pets.resize_with(index + 1, || None);
-                }
-                stable.stabled_pets[index] = Some(pet_info);
-            } else if slot == PetSaveMode::NotInSlot as i16 {
-                stable.unslotted_pets.push(pet_info);
-            } else {
-                continue;
-            }
-
-            loaded = loaded.saturating_add(1);
-        }
-
-        let selected_summoned_pet =
-            Pet::get_load_pet_info(&stable, 0, summoned_pet_number, None).is_some();
-        if selected_summoned_pet {
-            stable.current_pet_index = stable
-                .active_pets
-                .iter()
-                .position(|pet| {
-                    pet.as_ref()
-                        .is_some_and(|pet| pet.pet_number == summoned_pet_number)
-                })
-                .map(|index| index as u32);
-        }
-
-        if !self.update_player_pet_lifecycle_state_like_cpp(|state| {
-            // C++ remembers the selected controlled pet on the Player even
-            // while the live Pet object is absent from Map storage.
-            if selected_summoned_pet {
-                state.temporary_unsummoned_pet_number = summoned_pet_number;
-            }
-            state.stable = stable;
-            state.character_rows_empty_authority_complete = !query_had_rows;
-        }) {
-            return 0;
-        }
-        loaded
     }
 
     pub(crate) fn load_represented_pet_spell_rows_like_cpp(
@@ -25907,36 +25254,6 @@ impl WorldSession {
                 .insert(pet_number, effects);
         }
         loaded
-    }
-
-    pub(crate) fn load_represented_pet_declined_names_like_cpp(
-        &mut self,
-        pet_number: u32,
-        row: Option<CharacterPetDeclinedNamesRowLikeCpp>,
-    ) -> bool {
-        self.invalidate_represented_character_pet_empty_authority_like_cpp();
-        if let Some(row) = row {
-            self.pet_load_query_holder_rows_like_cpp
-                .declined_names
-                .insert(pet_number, row);
-            true
-        } else {
-            self.pet_load_query_holder_rows_like_cpp
-                .declined_names
-                .remove(&pet_number)
-                .is_some()
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_temporary_unsummoned_pet_number_like_cpp(&self) -> u32 {
-        self.player_pet_lifecycle_state_snapshot_like_cpp()
-            .map_or(0, |state| state.temporary_unsummoned_pet_number)
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_pet_guid_like_cpp(&self) -> Option<ObjectGuid> {
-        self.player_pet_guid_state_like_cpp().flatten()
     }
 
     #[cfg(test)]
@@ -27570,435 +26887,6 @@ impl WorldSession {
             .is_some_and(|state| state.represented_status == Some(4))
     }
 
-    fn unsummon_represented_pet_temporary_if_any_like_cpp(&mut self) {
-        let Some(pet_guid) = self.player_pet_guid_state_like_cpp().flatten() else {
-            return;
-        };
-        let Some(pet_lifecycle) = self.player_pet_lifecycle_state_snapshot_like_cpp() else {
-            return;
-        };
-
-        self.request_temporary_pet_unsummon_like_cpp();
-
-        if let Some(manager) = self.canonical_map_manager.as_ref().map(Arc::clone)
-            && let Ok(mut manager) = manager.lock()
-        {
-            let mut removed = false;
-            let mut temporary_pet_number = None;
-            let mut temporary_pet_created_by_spell = 0;
-            manager.do_for_all_maps_mut(|managed| {
-                if removed {
-                    return;
-                }
-                let Some(pet) = managed.map().get_typed_pet(pet_guid) else {
-                    return;
-                };
-                if pet_lifecycle.temporary_unsummoned_pet_number == 0
-                    && pet.is_controlled()
-                    && !pet.is_temporary_summoned()
-                {
-                    temporary_pet_number = pet
-                        .creature()
-                        .unit()
-                        .subsystems()
-                        .control
-                        .charm_info
-                        .as_ref()
-                        .map(|charm_info| charm_info.pet_number);
-                    temporary_pet_created_by_spell = pet.created_by_spell_id_like_cpp();
-                }
-                match managed.map_mut().remove_from_map_like_cpp(pet_guid, false) {
-                    Ok(_) => removed = true,
-                    Err(wow_map::RemoveFromMapError::ObjectNotFound { .. }) => {}
-                    Err(_) => {}
-                }
-            });
-            if let Some(pet_number) = temporary_pet_number.filter(|pet_number| *pet_number != 0) {
-                let _ = self.update_player_pet_lifecycle_state_like_cpp(|state| {
-                    state.temporary_unsummoned_pet_number = pet_number;
-                    state.old_pet_spell = temporary_pet_created_by_spell;
-                });
-            }
-        }
-
-        let _ = self.set_player_pet_guid_like_cpp(None);
-        #[cfg(test)]
-        {
-            self.represented_pet_created_by_spell_like_cpp = 0;
-            self.represented_pet_react_state_like_cpp =
-                wow_packet::packets::pet::REACT_DEFENSIVE_LIKE_CPP;
-            self.represented_pet_command_state_like_cpp =
-                wow_packet::packets::pet::COMMAND_FOLLOW_LIKE_CPP;
-        }
-        let _ = self.update_player_pet_lifecycle_state_like_cpp(|state| {
-            state.temporary_mount_react_state = None;
-        });
-    }
-
-    fn represented_pet_stable_info_by_number_like_cpp(
-        &self,
-        pet_number: u32,
-    ) -> Option<PetStableInfo> {
-        let pet_lifecycle = self.player_pet_lifecycle_state_snapshot_like_cpp()?;
-        pet_lifecycle
-            .stable
-            .active_pets
-            .iter()
-            .chain(pet_lifecycle.stable.stabled_pets.iter())
-            .flatten()
-            .find(|pet| pet.pet_number == pet_number)
-            .cloned()
-            .or_else(|| {
-                pet_lifecycle
-                    .stable
-                    .unslotted_pets
-                    .iter()
-                    .find(|pet| pet.pet_number == pet_number)
-                    .cloned()
-            })
-    }
-
-    fn is_pet_need_be_temporary_unsummoned_like_cpp(&self) -> bool {
-        !self.player_is_in_world_for_registry_like_cpp()
-            || self.resolved_player_is_alive_like_cpp() != Some(true)
-            || self
-                .resolved_player_movement_flags_like_cpp()
-                .is_none_or(|flags| flags.contains(MovementFlag::FLYING))
-    }
-
-    fn validate_represented_pet_action_bar_like_cpp(
-        &self,
-        charm_info: &mut wow_entities::CharmInfoState,
-    ) {
-        let Some(spell_store) = self.spell_store() else {
-            return;
-        };
-
-        for button in &mut charm_info.action_bar {
-            // C++ `UNIT_ACTION_BUTTON_TYPE` drops the low bit after `MAKE_UNIT_ACTION_BUTTON`
-            // stores `ActiveStates << 23`; recover the just-loaded type to apply the
-            // intended `LoadPetActionBar` validation without changing the packed wire shape.
-            let action_type = ((*button >> 23) & 0xFF) as u8;
-            if !matches!(
-                action_type,
-                wow_entities::ACT_DISABLED_LIKE_CPP
-                    | wow_entities::ACT_ENABLED_LIKE_CPP
-                    | wow_entities::ACT_PASSIVE_LIKE_CPP
-            ) {
-                continue;
-            }
-
-            let action = wow_entities::unit_action_button_action_like_cpp(*button);
-            if spell_store
-                .get(i32::try_from(action).unwrap_or(i32::MAX))
-                .is_none()
-            {
-                *button = wow_entities::make_unit_action_button_like_cpp(
-                    0,
-                    wow_entities::ACT_PASSIVE_LIKE_CPP,
-                );
-                continue;
-            }
-
-            if self
-                .spell_misc_store()
-                .is_some_and(|store| !store.is_autocastable_like_cpp(action))
-            {
-                *button = wow_entities::make_unit_action_button_like_cpp(
-                    action,
-                    wow_entities::ACT_PASSIVE_LIKE_CPP,
-                );
-            }
-        }
-    }
-
-    pub(crate) fn resummon_pet_temporary_unsummoned_like_cpp(&mut self) {
-        #[cfg(test)]
-        {
-            self.temporary_pet_resummon_requests_like_cpp = self
-                .temporary_pet_resummon_requests_like_cpp
-                .saturating_add(1);
-        }
-
-        let Some(pet_lifecycle) = self.player_pet_lifecycle_state_snapshot_like_cpp() else {
-            return;
-        };
-        let pet_number = pet_lifecycle.temporary_unsummoned_pet_number;
-        if pet_number == 0
-            || self.is_pet_need_be_temporary_unsummoned_like_cpp()
-            || self.player_pet_guid_state_like_cpp().flatten().is_some()
-        {
-            return;
-        }
-
-        self.invalidate_represented_character_pet_empty_authority_like_cpp();
-
-        let load_info = Pet::get_load_pet_info(&pet_lifecycle.stable, 0, pet_number, None);
-        let stable_info = load_info
-            .filter(|info| info.pet_number == pet_number)
-            .and_then(|_| self.represented_pet_stable_info_by_number_like_cpp(pet_number));
-        let inserted_guid = stable_info.and_then(|info| {
-            let owner_guid = self.player_guid()?;
-            let map_id = u32::from(self.player_map_id_like_cpp());
-            let instance_id = self
-                .current_canonical_player_map_key_like_cpp()
-                .map(|key| key.instance_id)?;
-            let position = self.player_position_like_cpp()?;
-            let creature_id = info.creature_id;
-            let pet_guid = ObjectGuid::create_world_object(
-                HighGuid::Pet,
-                0,
-                1,
-                self.player_map_id_like_cpp(),
-                instance_id,
-                creature_id,
-                i64::from(pet_number),
-            );
-
-            let mut pet = Pet::new(owner_guid, info.pet_type);
-            pet.set_created_by_spell_id_like_cpp(info.created_by_spell_id);
-            pet.set_specialization(info.specialization_id);
-            pet.creature_mut()
-                .unit_mut()
-                .world_mut()
-                .object_mut()
-                .create(pet_guid);
-            pet.creature_mut()
-                .unit_mut()
-                .world_mut()
-                .object_mut()
-                .set_entry(creature_id);
-            let _ = pet
-                .creature_mut()
-                .unit_mut()
-                .world_mut()
-                .set_map(map_id, instance_id);
-            pet.creature_mut().unit_mut().world_mut().relocate(position);
-            if !info.name.is_empty() {
-                pet.creature_mut()
-                    .unit_mut()
-                    .world_mut()
-                    .set_name(info.name);
-            }
-            if info.display_id != 0 {
-                pet.creature_mut()
-                    .set_display_id(info.display_id, true, None);
-            }
-            if info.level != 0 {
-                pet.creature_mut().unit_mut().set_level(info.level);
-            }
-            pet.set_pet_experience(info.experience);
-            if let Some(next_level_experience) = self
-                .resolved_player_xp_for_level_like_cpp(info.level)
-                .map(Pet::pet_next_level_xp_for_owner_level)
-            {
-                pet.set_pet_next_level_experience(next_level_experience);
-            }
-            if info.health == 0 && info.pet_type == PetType::Hunter {
-                pet.creature_mut()
-                    .unit_mut()
-                    .set_death_state(wow_constants::DeathState::JustDied);
-                pet.creature_mut().unit_mut().set_health(0);
-            } else if info.health != 0 {
-                pet.creature_mut()
-                    .unit_mut()
-                    .set_max_health(u64::from(info.health));
-                pet.creature_mut()
-                    .unit_mut()
-                    .set_health(u64::from(info.health));
-            }
-            if info.mana != 0 {
-                pet.creature_mut()
-                    .unit_mut()
-                    .set_power(PowerType::Mana, info.mana as i32);
-            }
-            pet.creature_mut().set_react_state(info.react_state);
-            let charm_info = pet
-                .creature_mut()
-                .unit_mut()
-                .subsystems_mut()
-                .control
-                .init_charm_info();
-            charm_info.pet_number = pet_number;
-            charm_info.command_state = wow_packet::packets::pet::COMMAND_FOLLOW_LIKE_CPP;
-            charm_info.load_pet_action_bar_like_cpp(&info.action_bar);
-            self.validate_represented_pet_action_bar_like_cpp(charm_info);
-            if let Some(spells) = self
-                .pet_load_query_holder_rows_like_cpp
-                .spells
-                .get(&pet_number)
-            {
-                for spell in spells {
-                    pet.add_spell(
-                        spell.spell_id,
-                        active_state_from_db_like_cpp(spell.active),
-                        PetSpellState::Unchanged,
-                        PetSpellType::Normal,
-                    );
-                }
-            }
-            let spell_history = &mut pet
-                .creature_mut()
-                .unit_mut()
-                .subsystems_mut()
-                .spells
-                .history;
-            if let Some(cooldowns) = self
-                .pet_load_query_holder_rows_like_cpp
-                .spell_cooldowns
-                .get(&pet_number)
-            {
-                for cooldown in cooldowns {
-                    spell_history.add_cooldown(
-                        cooldown.spell_id,
-                        0,
-                        unix_secs_to_ms_like_cpp(cooldown.cooldown_end_unix_secs),
-                        cooldown.category_id,
-                        unix_secs_to_ms_like_cpp(cooldown.category_end_unix_secs),
-                        false,
-                    );
-                }
-            }
-            if let Some(charges) = self
-                .pet_load_query_holder_rows_like_cpp
-                .spell_charges
-                .get(&pet_number)
-            {
-                for charge in charges {
-                    spell_history.add_charge_state_like_cpp(
-                        charge.category_id,
-                        unix_secs_to_ms_like_cpp(charge.recharge_start_unix_secs),
-                        unix_secs_to_ms_like_cpp(charge.recharge_end_unix_secs),
-                    );
-                }
-            }
-            let pet_aura_effects = self
-                .pet_load_query_holder_rows_like_cpp
-                .aura_effects
-                .get(&pet_number)
-                .cloned()
-                .unwrap_or_default();
-            if let Some(auras) = self
-                .pet_load_query_holder_rows_like_cpp
-                .auras
-                .get(&pet_number)
-            {
-                let aura_subsystem = &mut pet.creature_mut().unit_mut().subsystems_mut().auras;
-                for (index, aura) in auras.iter().enumerate() {
-                    let Some(slot) = represented_pet_aura_slot_like_cpp(index) else {
-                        break;
-                    };
-                    let caster_guid = if aura.caster_guid.is_empty() {
-                        pet_guid
-                    } else {
-                        aura.caster_guid
-                    };
-                    let applied = wow_entities::AppliedAuraRef::new(
-                        aura.spell_id,
-                        caster_guid,
-                        slot,
-                        aura.effect_mask,
-                    );
-                    let aura_ref = wow_entities::AuraRef::new(aura.spell_id, caster_guid);
-                    aura_subsystem.add_owned(wow_entities::OwnedAuraRef::new(
-                        aura.spell_id,
-                        caster_guid,
-                        None,
-                    ));
-                    aura_subsystem.add_applied(applied);
-                    aura_subsystem.set_loaded_aura_state_like_cpp(
-                        aura_ref,
-                        wow_entities::LoadedAuraStateLikeCpp::new(
-                            aura.max_duration_ms,
-                            aura.remain_time_ms,
-                            aura.remain_charges,
-                            aura.stack_count,
-                            aura.recalculate_mask,
-                        ),
-                    );
-                    aura_subsystem.visible_auras.insert(slot, aura_ref);
-
-                    let effect_amounts: Vec<_> = pet_aura_effects
-                        .iter()
-                        .filter(|effect| {
-                            let effect_caster_guid = if effect.caster_guid.is_empty() {
-                                pet_guid
-                            } else {
-                                effect.caster_guid
-                            };
-                            effect_caster_guid == caster_guid
-                                && effect.spell_id == aura.spell_id
-                                && effect.effect_mask == aura.effect_mask
-                        })
-                        .map(|effect| {
-                            let effect_ref = wow_entities::AppliedAuraRef::new(
-                                aura.spell_id,
-                                caster_guid,
-                                slot,
-                                1u32 << u32::from(effect.effect_index),
-                            );
-                            aura_subsystem
-                                .applied_aura_amounts
-                                .insert(effect_ref, effect.amount);
-                            wow_entities::VisibleAuraEffectAmountLikeCpp {
-                                effect_index: effect.effect_index,
-                                amount: effect.amount,
-                            }
-                        })
-                        .collect();
-                    aura_subsystem.visible_aura_applications_like_cpp.insert(
-                        slot,
-                        wow_entities::VisibleAuraApplicationLikeCpp::new(
-                            aura.effect_mask,
-                            effect_amounts,
-                        ),
-                    );
-                }
-            }
-            if info.pet_type == PetType::Hunter
-                && let Some(declined_names) = self
-                    .pet_load_query_holder_rows_like_cpp
-                    .declined_names
-                    .get(&pet_number)
-            {
-                pet.set_declined_names(Some(PetDeclinedNamesLikeCpp {
-                    names: declined_names.names.clone(),
-                }));
-            }
-
-            let manager = self.canonical_map_manager.as_ref().map(Arc::clone)?;
-            let mut manager = manager.lock().ok()?;
-            if manager.find_map_mut(map_id, instance_id).is_none() {
-                manager.create_world_map(map_id, instance_id);
-            }
-            let managed = manager.find_map_mut(map_id, instance_id)?;
-            managed
-                .map_mut()
-                .insert_map_object_record(wow_entities::MapObjectRecord::new_pet(pet).ok()?)
-                .ok()?;
-            Some(pet_guid)
-        });
-
-        if !self.update_player_pet_lifecycle_state_like_cpp(|state| {
-            state.temporary_unsummoned_pet_number = 0;
-        }) {
-            return;
-        }
-        if let Some(pet_guid) = inserted_guid {
-            let _ = self.set_player_pet_guid_like_cpp(Some(pet_guid));
-            #[cfg(test)]
-            if let Some(info) = self.represented_pet_stable_info_by_number_like_cpp(pet_number) {
-                self.represented_pet_created_by_spell_like_cpp = info.created_by_spell_id;
-                self.represented_pet_react_state_like_cpp = info.react_state as u8;
-            }
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn resummon_pet_temporary_unsummoned_if_any_like_cpp(&mut self) {
-        self.resummon_pet_temporary_unsummoned_like_cpp();
-    }
-
     #[cfg_attr(not(test), allow(unused_variables))]
     pub(crate) fn battlemaster_hello_like_cpp(&mut self, unit: ObjectGuid) -> bool {
         let Some((npc_flags, entry)) =
@@ -29610,63 +28498,6 @@ impl WorldSession {
             .map_or(0, |state| state.money)
     }
 
-    #[cfg_attr(not(test), allow(unused_variables))]
-    pub(crate) fn record_represented_sign_petition_like_cpp(
-        &mut self,
-        petition_guid: ObjectGuid,
-        choice: u8,
-    ) {
-        #[cfg(test)]
-        self.represented_sign_petitions_like_cpp
-            .push(RepresentedSignPetitionLikeCpp {
-                petition_guid,
-                choice,
-            });
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_sign_petitions_like_cpp(&self) -> &[RepresentedSignPetitionLikeCpp] {
-        &self.represented_sign_petitions_like_cpp
-    }
-
-    #[cfg_attr(not(test), allow(unused_variables))]
-    pub(crate) fn record_represented_decline_petition_like_cpp(
-        &mut self,
-        petition_guid: ObjectGuid,
-    ) {
-        #[cfg(test)]
-        self.represented_decline_petitions_like_cpp
-            .push(RepresentedDeclinePetitionLikeCpp { petition_guid });
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_decline_petitions_like_cpp(
-        &self,
-    ) -> &[RepresentedDeclinePetitionLikeCpp] {
-        &self.represented_decline_petitions_like_cpp
-    }
-
-    #[cfg_attr(not(test), allow(unused_variables))]
-    pub(crate) fn record_represented_query_petition_like_cpp(
-        &mut self,
-        petition_id: u32,
-        item_guid: ObjectGuid,
-    ) {
-        #[cfg(test)]
-        self.represented_query_petitions_like_cpp
-            .push(RepresentedQueryPetitionLikeCpp {
-                petition_id,
-                item_guid,
-            });
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_query_petitions_like_cpp(
-        &self,
-    ) -> &[RepresentedQueryPetitionLikeCpp] {
-        &self.represented_query_petitions_like_cpp
-    }
-
     pub(crate) fn record_represented_silence_party_talker_like_cpp(
         &mut self,
         target: ObjectGuid,
@@ -30340,34 +29171,6 @@ impl WorldSession {
             .is_some()
     }
 
-    #[cfg(test)]
-    fn record_represented_tapper_pet_killed_unit_hooks_like_cpp(
-        &mut self,
-        creature_guid: ObjectGuid,
-    ) {
-        let (Some(player_guid), Some(pet_guid)) = (
-            self.player_guid(),
-            self.player_pet_guid_state_like_cpp().flatten(),
-        ) else {
-            return;
-        };
-        let tapper_has_current_player = self
-            .mutate_world_creature(creature_guid, |creature| {
-                creature.creature.tap_list().contains(&player_guid)
-            })
-            .unwrap_or(true);
-        if !tapper_has_current_player {
-            return;
-        }
-        self.represented_creature_kill_events_like_cpp.push(
-            RepresentedCreatureKillEventLikeCpp::TapperPetKilledUnitAi {
-                tapper_guid: player_guid,
-                pet_guid,
-                victim_guid: creature_guid,
-            },
-        );
-    }
-
     pub(crate) fn represented_current_vehicle_seat_can_switch_from_like_cpp(&self) -> bool {
         self.player_vehicle_seat_state_like_cpp()
             .and_then(|(flags, _)| flags)
@@ -30623,742 +29426,6 @@ impl WorldSession {
         true
     }
 
-    /// Test/setup seam for represented `BattlePetMgr::_pets`.
-    #[cfg(test)]
-    pub(crate) fn add_represented_battle_pet_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-        flags: u16,
-        save_info: RepresentedBattlePetSaveInfoLikeCpp,
-    ) {
-        self.represented_battle_pets_like_cpp.insert(
-            pet_guid,
-            RepresentedBattlePetDataLikeCpp::minimal_like_cpp(flags, save_info),
-        );
-    }
-
-    /// Test/setup seam for represented `BattlePet::PacketInfo` rows already
-    /// loaded into `BattlePetMgr::_pets`.
-    #[cfg(test)]
-    pub(crate) fn add_represented_battle_pet_packet_info_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-        packet_info: RepresentedBattlePetDataLikeCpp,
-    ) {
-        self.represented_battle_pets_like_cpp
-            .insert(pet_guid, packet_info);
-    }
-
-    /// C++ `BattlePetMgr::HasJournalLock`.
-    pub(crate) fn has_represented_battle_pet_journal_lock_like_cpp(&self) -> bool {
-        if let Some(attachment) = &self.battle_pet_account_attachment_like_cpp {
-            return attachment.has_lease_like_cpp();
-        }
-        #[cfg(test)]
-        return self.represented_battle_pet_journal_lock_like_cpp;
-        #[cfg(not(test))]
-        false
-    }
-
-    /// C++ `BattlePetMgr::SendJournalLockStatus`, represented as the successful
-    /// local acquisition path until the global world journal-lock owner exists.
-    pub(crate) async fn send_battle_pet_journal_lock_status_like_cpp(&mut self) {
-        if let Some(attachment) = &self.battle_pet_account_attachment_like_cpp {
-            let acquired = attachment.try_acquire_lease_like_cpp().await;
-            if acquired {
-                self.send_packet_realm(&wow_packet::packets::misc::BattlePetJournalLockAcquired);
-            } else {
-                self.send_packet_realm(&wow_packet::packets::misc::BattlePetJournalLockDenied);
-            }
-            return;
-        }
-        #[cfg(test)]
-        {
-            self.represented_battle_pet_journal_lock_like_cpp = true;
-            self.send_packet_realm(&wow_packet::packets::misc::BattlePetJournalLockAcquired);
-        }
-        #[cfg(not(test))]
-        self.send_packet_realm(&wow_packet::packets::misc::BattlePetJournalLockDenied);
-    }
-
-    /// C++ `BattlePetMgr::ClearFanfare`.
-    #[cfg(test)]
-    pub(crate) fn battle_pet_clear_fanfare_like_cpp(&mut self, pet_guid: ObjectGuid) -> bool {
-        let Some(pet) = self.represented_battle_pets_like_cpp.get_mut(&pet_guid) else {
-            return false;
-        };
-
-        pet.flags &= !BATTLE_PET_FLAG_FANFARE_NEEDED_LIKE_CPP;
-        if pet.save_info != RepresentedBattlePetSaveInfoLikeCpp::New {
-            pet.save_info = RepresentedBattlePetSaveInfoLikeCpp::Changed;
-        }
-        true
-    }
-
-    pub(crate) async fn battle_pet_clear_fanfare_durable_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-    ) -> bool {
-        let Some(attachment) = &self.battle_pet_account_attachment_like_cpp else {
-            #[cfg(test)]
-            return self.battle_pet_clear_fanfare_like_cpp(pet_guid);
-            #[cfg(not(test))]
-            return false;
-        };
-        let owner = Arc::clone(attachment.owner_like_cpp());
-        match owner
-            .try_mutate_pet_without_lease_like_cpp(pet_guid, |pet| {
-                pet.flags &= !BATTLE_PET_FLAG_FANFARE_NEEDED_LIKE_CPP;
-            })
-            .await
-        {
-            Ok(_) => true,
-            Err(error) => {
-                self.log_battle_pet_mutation_failure_like_cpp("clear fanfare", pet_guid, &error);
-                false
-            }
-        }
-    }
-
-    /// C++ `BattlePetMgr::RemovePet`.
-    #[cfg(test)]
-    pub(crate) fn battle_pet_remove_pet_like_cpp(&mut self, pet_guid: ObjectGuid) -> bool {
-        if !self.has_represented_battle_pet_journal_lock_like_cpp() {
-            return false;
-        }
-
-        let Some(pet) = self.represented_battle_pets_like_cpp.get_mut(&pet_guid) else {
-            return false;
-        };
-
-        pet.save_info = RepresentedBattlePetSaveInfoLikeCpp::Removed;
-        true
-    }
-
-    pub(crate) async fn battle_pet_remove_pet_durable_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-    ) -> bool {
-        let Some(attachment) = &self.battle_pet_account_attachment_like_cpp else {
-            #[cfg(test)]
-            return self.battle_pet_remove_pet_like_cpp(pet_guid);
-            #[cfg(not(test))]
-            return false;
-        };
-        let owner = Arc::clone(attachment.owner_like_cpp());
-        let lease = attachment.lease_id_like_cpp();
-        match owner.try_remove_pet_like_cpp(lease, pet_guid).await {
-            Ok(()) => true,
-            Err(error) => {
-                self.log_battle_pet_mutation_failure_like_cpp("remove", pet_guid, &error);
-                false
-            }
-        }
-    }
-
-    /// C++ `BattlePetMgr::CageBattlePet`, represented at the battle-pet state
-    /// boundary. Inventory placement is still external: the caller must pass
-    /// the already-resolved `CanStoreNewItem`/`StoreNewItem` outcomes.
-    #[cfg(test)]
-    pub(crate) fn battle_pet_cage_battle_pet_represented_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-        inventory_can_store: bool,
-        item_stored: bool,
-    ) -> RepresentedBattlePetCageOutcomeLikeCpp {
-        if !self.has_represented_battle_pet_journal_lock_like_cpp() {
-            return RepresentedBattlePetCageOutcomeLikeCpp::NoJournalLock;
-        }
-
-        let Some(pet) = self
-            .represented_battle_pets_like_cpp
-            .get(&pet_guid)
-            .cloned()
-        else {
-            return RepresentedBattlePetCageOutcomeLikeCpp::UnknownPet;
-        };
-
-        if self.battle_pet_species_has_flag_like_cpp(
-            pet.species,
-            wow_data::BATTLE_PET_SPECIES_FLAG_NOT_TRADABLE_LIKE_CPP,
-        ) {
-            return RepresentedBattlePetCageOutcomeLikeCpp::NotTradable;
-        }
-
-        if self
-            .represented_battle_pet_slots_like_cpp
-            .iter()
-            .any(|slot| slot.pet_guid == Some(pet_guid))
-        {
-            return RepresentedBattlePetCageOutcomeLikeCpp::InBattleSlot;
-        }
-
-        if pet.health < pet.max_health {
-            return RepresentedBattlePetCageOutcomeLikeCpp::Damaged;
-        }
-
-        if !inventory_can_store {
-            return RepresentedBattlePetCageOutcomeLikeCpp::InventoryUnavailable;
-        }
-
-        if !item_stored {
-            return RepresentedBattlePetCageOutcomeLikeCpp::StoreFailed;
-        }
-
-        let cage_item = RepresentedBattlePetCageItemLikeCpp {
-            item_id: BATTLE_PET_CAGE_ITEM_ID_LIKE_CPP,
-            species_id: pet.species,
-            breed_data: u32::from(pet.breed) | (u32::from(pet.quality) << 24),
-            level: pet.level,
-            display_id: pet.display_id,
-        };
-        #[cfg(test)]
-        self.represented_battle_pet_cage_items_like_cpp
-            .push(cage_item);
-        #[cfg(not(test))]
-        let _ = cage_item;
-
-        let _ = self.battle_pet_remove_pet_like_cpp(pet_guid);
-        self.send_packet(&wow_packet::packets::misc::BattlePetDeleted { pet_guid });
-
-        if self.represented_summoned_battle_pet_guid_like_cpp() == Some(pet_guid) {
-            let _cleared = self.mutate_canonical_player_like_cpp(|player| {
-                player.clear_battle_pet_data_like_cpp();
-            });
-            #[cfg(test)]
-            if _cleared.is_none() {
-                self.represented_summoned_battle_pet_guid_like_cpp = None;
-            }
-        }
-
-        RepresentedBattlePetCageOutcomeLikeCpp::Caged(cage_item)
-    }
-
-    /// C++ `BattlePetMgr::ModifyName`, represented without live summoned-creature
-    /// timestamp propagation until the companion creature runtime exists.
-    #[cfg(test)]
-    pub(crate) fn battle_pet_modify_name_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-        name: String,
-        declined_names: Option<wow_packet::packets::misc::DeclinedNamesLikeCpp>,
-        timestamp: i64,
-    ) -> bool {
-        if !self.has_represented_battle_pet_journal_lock_like_cpp() {
-            return false;
-        }
-
-        let Some(pet) = self.represented_battle_pets_like_cpp.get_mut(&pet_guid) else {
-            return false;
-        };
-
-        pet.name = name;
-        pet.name_timestamp = timestamp;
-        pet.declined_names = declined_names;
-        if pet.save_info != RepresentedBattlePetSaveInfoLikeCpp::New {
-            pet.save_info = RepresentedBattlePetSaveInfoLikeCpp::Changed;
-        }
-        true
-    }
-
-    pub(crate) async fn battle_pet_modify_name_durable_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-        name: String,
-        declined_names: Option<wow_packet::packets::misc::DeclinedNamesLikeCpp>,
-        timestamp: i64,
-    ) -> bool {
-        let Some(attachment) = &self.battle_pet_account_attachment_like_cpp else {
-            #[cfg(test)]
-            return self.battle_pet_modify_name_like_cpp(pet_guid, name, declined_names, timestamp);
-            #[cfg(not(test))]
-            return false;
-        };
-        let owner = Arc::clone(attachment.owner_like_cpp());
-        let lease = attachment.lease_id_like_cpp();
-        match owner
-            .try_mutate_pet_like_cpp(lease, pet_guid, move |pet| {
-                pet.name = name;
-                pet.name_timestamp = timestamp;
-                pet.declined_names = declined_names;
-            })
-            .await
-        {
-            Ok(_) => true,
-            Err(error) => {
-                self.log_battle_pet_mutation_failure_like_cpp("modify name", pet_guid, &error);
-                false
-            }
-        }
-    }
-
-    /// C++ `WorldSession::HandleBattlePetSetFlags` flag mutation.
-    #[cfg(test)]
-    pub(crate) fn battle_pet_set_flags_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-        flags: u16,
-        control_type: u8,
-    ) -> bool {
-        let Some(pet) = self.represented_battle_pets_like_cpp.get_mut(&pet_guid) else {
-            return false;
-        };
-
-        if control_type == BATTLE_PET_FLAGS_CONTROL_TYPE_APPLY_LIKE_CPP {
-            pet.flags |= flags;
-        } else {
-            pet.flags &= !flags;
-        }
-
-        if pet.save_info != RepresentedBattlePetSaveInfoLikeCpp::New {
-            pet.save_info = RepresentedBattlePetSaveInfoLikeCpp::Changed;
-        }
-        true
-    }
-
-    pub(crate) async fn battle_pet_set_flags_durable_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-        flags: u16,
-        control_type: u8,
-    ) -> bool {
-        let Some(attachment) = &self.battle_pet_account_attachment_like_cpp else {
-            #[cfg(test)]
-            return self.battle_pet_set_flags_like_cpp(pet_guid, flags, control_type);
-            #[cfg(not(test))]
-            return false;
-        };
-        let owner = Arc::clone(attachment.owner_like_cpp());
-        let lease = attachment.lease_id_like_cpp();
-        match owner
-            .try_mutate_pet_like_cpp(lease, pet_guid, move |pet| {
-                if control_type == BATTLE_PET_FLAGS_CONTROL_TYPE_APPLY_LIKE_CPP {
-                    pet.flags |= flags;
-                } else {
-                    pet.flags &= !flags;
-                }
-            })
-            .await
-        {
-            Ok(_) => true,
-            Err(error) => {
-                self.log_battle_pet_mutation_failure_like_cpp("set flags", pet_guid, &error);
-                false
-            }
-        }
-    }
-
-    /// C++ `WorldSession::HandleBattlePetSetBattleSlot`.
-    #[cfg(test)]
-    pub(crate) fn battle_pet_set_battle_slot_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-        slot: u8,
-    ) -> bool {
-        if !self
-            .represented_battle_pets_like_cpp
-            .contains_key(&pet_guid)
-        {
-            return false;
-        }
-
-        let Some(slot_ref) = self
-            .represented_battle_pet_slots_like_cpp
-            .get_mut(slot as usize)
-        else {
-            return false;
-        };
-        slot_ref.pet_guid = Some(pet_guid);
-        self.invalidate_canonical_player_spell_hit_aura_authority_like_cpp();
-        true
-    }
-
-    pub(crate) async fn battle_pet_set_battle_slot_durable_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-        slot: u8,
-    ) -> bool {
-        let Some(attachment) = &self.battle_pet_account_attachment_like_cpp else {
-            #[cfg(test)]
-            return self.battle_pet_set_battle_slot_like_cpp(pet_guid, slot);
-            #[cfg(not(test))]
-            return false;
-        };
-        let owner = Arc::clone(attachment.owner_like_cpp());
-        let lease = attachment.lease_id_like_cpp();
-        match owner.try_set_slot_like_cpp(lease, pet_guid, slot).await {
-            Ok(_) => {
-                self.invalidate_canonical_player_spell_hit_aura_authority_like_cpp();
-                true
-            }
-            Err(error) => {
-                self.log_battle_pet_mutation_failure_like_cpp("set battle slot", pet_guid, &error);
-                false
-            }
-        }
-    }
-
-    fn log_battle_pet_mutation_failure_like_cpp(
-        &self,
-        operation: &'static str,
-        pet_guid: ObjectGuid,
-        error: &BattlePetMutationFailureLikeCpp,
-    ) {
-        if matches!(
-            error,
-            BattlePetMutationFailureLikeCpp::MissingAuthority
-                | BattlePetMutationFailureLikeCpp::JournalLocked
-                | BattlePetMutationFailureLikeCpp::UnknownPet
-        ) {
-            return;
-        }
-        warn!(
-            account = self.account_id,
-            ?pet_guid,
-            ?error,
-            operation,
-            "Durable battle-pet mutation failed"
-        );
-    }
-
-    /// C++ `BattlePetMgr::UnlockSlot`.
-    #[cfg(test)]
-    pub(crate) fn battle_pet_unlock_slot_like_cpp(&mut self, slot: u8) -> bool {
-        let Some(slot_ref) = self
-            .represented_battle_pet_slots_like_cpp
-            .get_mut(slot as usize)
-        else {
-            return false;
-        };
-
-        if !slot_ref.locked {
-            return false;
-        }
-
-        slot_ref.locked = false;
-        let packet_slot = slot_ref.packet_slot_like_cpp();
-        self.invalidate_canonical_player_spell_hit_aura_authority_like_cpp();
-        self.send_packet(&wow_packet::packets::misc::PetBattleSlotUpdates {
-            slots: vec![packet_slot],
-            auto_slotted: false,
-            new_slot: true,
-        });
-        true
-    }
-
-    /// C++ `BattlePetMgr::SendUpdates`.
-    #[cfg(test)]
-    pub(crate) fn send_battle_pet_updates_like_cpp(
-        &mut self,
-        pet_guids: &[ObjectGuid],
-        pet_added: bool,
-    ) -> usize {
-        let pets: Vec<_> = pet_guids
-            .iter()
-            .filter_map(|pet_guid| {
-                let pet = self.represented_battle_pets_like_cpp.get(pet_guid)?;
-                if pet.save_info == RepresentedBattlePetSaveInfoLikeCpp::Removed {
-                    return None;
-                }
-                Some(pet.packet_info_like_cpp(*pet_guid))
-            })
-            .collect();
-        let sent_count = pets.len();
-        self.send_packet(&wow_packet::packets::misc::BattlePetUpdates { pets, pet_added });
-        sent_count
-    }
-
-    /// C++ `BattlePetMgr::GetMaxPetLevel`.
-    pub(crate) fn battle_pet_max_pet_level_like_cpp(&self) -> Option<u16> {
-        if let Some(attachment) = &self.battle_pet_account_attachment_like_cpp {
-            return Some(attachment.owner_like_cpp().max_pet_level_like_cpp());
-        }
-        #[cfg(test)]
-        return Some(
-            self.represented_battle_pets_like_cpp
-                .values()
-                .filter(|pet| pet.save_info != RepresentedBattlePetSaveInfoLikeCpp::Removed)
-                .map(|pet| pet.level)
-                .max()
-                .unwrap_or(0),
-        );
-        #[cfg(not(test))]
-        None
-    }
-
-    /// C++ `BattlePetMgr::GetPetCount`.
-    #[cfg(test)]
-    pub(crate) fn battle_pet_count_like_cpp(
-        &self,
-        species: u32,
-        owner_guid: Option<ObjectGuid>,
-    ) -> u8 {
-        if let Some(attachment) = &self.battle_pet_account_attachment_like_cpp {
-            return attachment
-                .owner_like_cpp()
-                .pet_count_like_cpp(species, owner_guid);
-        }
-        let species_flags = self
-            .battle_pet_species_store
-            .as_ref()
-            .and_then(|store| store.get(species))
-            .map(|entry| entry.flags)
-            .unwrap_or(0);
-        let not_account_wide =
-            species_flags & wow_data::BATTLE_PET_SPECIES_FLAG_NOT_ACCOUNT_WIDE_LIKE_CPP != 0;
-
-        let count = self
-            .represented_battle_pets_like_cpp
-            .values()
-            .filter(|pet| pet.species == species)
-            .filter(|pet| pet.save_info != RepresentedBattlePetSaveInfoLikeCpp::Removed)
-            .filter(|pet| {
-                if not_account_wide {
-                    if let (Some(owner_guid), Some(owner_info)) = (owner_guid, pet.owner_info) {
-                        return owner_info.guid == owner_guid;
-                    }
-                }
-                true
-            })
-            .count();
-
-        u8::try_from(count).unwrap_or(u8::MAX)
-    }
-
-    /// C++ `BattlePetMgr::HasMaxPetCount`.
-    pub(crate) fn battle_pet_has_max_pet_count_like_cpp(
-        &self,
-        species: u32,
-        owner_guid: Option<ObjectGuid>,
-    ) -> Option<bool> {
-        if let Some(attachment) = &self.battle_pet_account_attachment_like_cpp {
-            return Some(
-                attachment
-                    .owner_like_cpp()
-                    .has_max_pet_count_like_cpp(species, owner_guid),
-            );
-        }
-        #[cfg(test)]
-        {
-            let Some(species_entry) = self
-                .battle_pet_species_store
-                .as_ref()
-                .and_then(|store| store.get(species))
-            else {
-                return Some(false);
-            };
-
-            let max_pets_per_species = if species_entry
-                .has_flag_like_cpp(wow_data::BATTLE_PET_SPECIES_FLAG_LEGACY_ACCOUNT_UNIQUE_LIKE_CPP)
-            {
-                1
-            } else {
-                DEFAULT_MAX_BATTLE_PETS_PER_SPECIES_LIKE_CPP
-            };
-
-            return Some(
-                self.battle_pet_count_like_cpp(species, owner_guid) >= max_pets_per_species,
-            );
-        }
-        #[cfg(not(test))]
-        None
-    }
-
-    #[cfg(test)]
-    fn next_represented_battle_pet_guid_like_cpp() -> ObjectGuid {
-        let counter = NEXT_REPRESENTED_BATTLE_PET_COUNTER_LIKE_CPP.fetch_add(1, Ordering::Relaxed);
-        ObjectGuid::create_global(HighGuid::BattlePet, 0, counter)
-    }
-
-    /// C++ `BattlePetMgr::AddPet`, represented without DB persistence and with
-    /// a local GUID counter until `sObjectMgr->GetGenerator<HighGuid::BattlePet>()`
-    /// is ported.
-    #[cfg(test)]
-    pub(crate) fn battle_pet_add_pet_represented_like_cpp(
-        &mut self,
-        species: u32,
-        display_id: u32,
-        breed: u16,
-        quality: u8,
-        level: u16,
-    ) -> Option<ObjectGuid> {
-        let species_entry = self
-            .battle_pet_species_store
-            .as_ref()
-            .and_then(|store| store.get(species))
-            .cloned()?;
-
-        if !species_entry.has_flag_like_cpp(wow_data::BATTLE_PET_SPECIES_FLAG_WELL_KNOWN_LIKE_CPP) {
-            return None;
-        }
-
-        let calculated_stats =
-            self.battle_pet_calculate_stats_like_cpp(breed, species, quality, level);
-        let pet_guid = Self::next_represented_battle_pet_guid_like_cpp();
-        let owner_info = if species_entry
-            .has_flag_like_cpp(wow_data::BATTLE_PET_SPECIES_FLAG_NOT_ACCOUNT_WIDE_LIKE_CPP)
-        {
-            self.player_guid().map(
-                |guid| wow_packet::packets::misc::BattlePetJournalPetOwnerInfo {
-                    guid,
-                    player_virtual_realm: 1,
-                    player_native_realm: 1,
-                },
-            )
-        } else {
-            None
-        };
-
-        let mut pet = RepresentedBattlePetDataLikeCpp {
-            species,
-            creature_id: u32::try_from(species_entry.creature_id).unwrap_or_default(),
-            display_id,
-            breed,
-            level,
-            exp: 0,
-            flags: 0,
-            power: 0,
-            health: 0,
-            max_health: 0,
-            speed: 0,
-            quality,
-            owner_info,
-            name: String::new(),
-            name_timestamp: 0,
-            declined_names: None,
-            save_info: RepresentedBattlePetSaveInfoLikeCpp::New,
-        };
-        Self::apply_battle_pet_calculated_stats_like_cpp(&mut pet, calculated_stats);
-
-        self.represented_battle_pets_like_cpp.insert(pet_guid, pet);
-        self.send_battle_pet_updates_like_cpp(&[pet_guid], true);
-        #[cfg(test)]
-        {
-            self.represented_battle_pet_unique_owned_criteria_like_cpp = self
-                .represented_battle_pet_unique_owned_criteria_like_cpp
-                .saturating_add(1);
-            self.represented_battle_pet_learned_new_pet_criteria_like_cpp
-                .push(species);
-        }
-
-        Some(pet_guid)
-    }
-
-    /// Durable account-scoped `BattlePetMgr::AddPet` replacement. Unlike the
-    /// target C++ split `HasMaxPetCount`/`AddPet` sequence, the owner rechecks
-    /// lease and capacity while reserving the insert. Publication happens only
-    /// after the Login DB pet row and idempotency receipt commit together.
-    pub(crate) async fn battle_pet_try_add_pet_durable_like_cpp(
-        &mut self,
-        request_key: [u8; 16],
-        species: u32,
-        display_id: u32,
-        breed: u16,
-        quality: u8,
-        level: u16,
-    ) -> Result<ObjectGuid, BattlePetAddFailureLikeCpp> {
-        let Some(attachment) = &self.battle_pet_account_attachment_like_cpp else {
-            #[cfg(test)]
-            return self
-                .battle_pet_add_pet_represented_like_cpp(species, display_id, breed, quality, level)
-                .ok_or(BattlePetAddFailureLikeCpp::InvalidSpecies);
-            #[cfg(not(test))]
-            return Err(BattlePetAddFailureLikeCpp::MissingAuthority);
-        };
-        let owner = Arc::clone(attachment.owner_like_cpp());
-        let lease_id = attachment.lease_id_like_cpp();
-        let outcome = owner
-            .try_add_pet_like_cpp(
-                lease_id,
-                BattlePetAddRequestLikeCpp {
-                    request_key: BattlePetAddRequestKeyLikeCpp::from_bytes(request_key),
-                    species,
-                    display_id,
-                    breed,
-                    quality,
-                    level,
-                    owner_guid: self.player_guid(),
-                },
-            )
-            .await?;
-        match outcome {
-            BattlePetAddOutcomeLikeCpp::Added(pet) => {
-                let guid = pet.guid;
-                self.send_packet(&wow_packet::packets::misc::BattlePetUpdates {
-                    pets: vec![pet],
-                    pet_added: true,
-                });
-                #[cfg(test)]
-                {
-                    self.represented_battle_pet_unique_owned_criteria_like_cpp = self
-                        .represented_battle_pet_unique_owned_criteria_like_cpp
-                        .saturating_add(1);
-                    self.represented_battle_pet_learned_new_pet_criteria_like_cpp
-                        .push(species);
-                }
-                Ok(guid)
-            }
-            BattlePetAddOutcomeLikeCpp::Replayed(pet) => Ok(pet.guid),
-        }
-    }
-
-    /// Publish one durable battle-pet addition exactly like the #160 session
-    /// seam (`SMSG_BATTLE_PET_UPDATES` with `pet_added` plus the C++
-    /// packet). The issue #161 saga calls this only after the pet and receipt
-    /// are durable. Packet recovery may call it again, so it deliberately has
-    /// no criteria side effects.
-    pub(crate) fn publish_battle_pet_trainer_purchase_add_like_cpp(
-        &mut self,
-        pet: wow_packet::packets::misc::BattlePetJournalPet,
-    ) -> bool {
-        let packet_enqueued = self
-            .send_tx()
-            .send(wow_packet::ServerPacket::to_bytes(
-                &wow_packet::packets::misc::BattlePetUpdates {
-                    pets: vec![pet],
-                    pet_added: true,
-                },
-            ))
-            .is_ok();
-        if !packet_enqueued {
-            warn!("Send channel closed for account {}", self.account_id);
-        }
-        packet_enqueued
-    }
-
-    /// Record the C++ `BattlePetMgr::AddPet` criteria hooks from durable
-    /// current state. Both C++ criteria are set-like (`UniquePetsOwned` uses
-    /// the current unique-species count and `LearnedNewPet` sets one species
-    /// to 1), so receipt recovery and packet re-sends are idempotent.
-    pub(crate) fn record_battle_pet_trainer_purchase_criteria_like_cpp(&mut self, species: u32) {
-        #[cfg(not(test))]
-        let _ = species;
-        #[cfg(test)]
-        {
-            self.represented_battle_pet_unique_owned_criteria_like_cpp = self
-                .battle_pet_account_attachment_like_cpp
-                .as_ref()
-                .map(|attachment| attachment.owner_like_cpp().unique_species_count_like_cpp())
-                .unwrap_or_else(|| {
-                    u32::try_from(
-                        self.represented_battle_pets_like_cpp
-                            .values()
-                            .map(|pet| pet.species)
-                            .collect::<BTreeSet<_>>()
-                            .len(),
-                    )
-                    .unwrap_or(u32::MAX)
-                });
-            if !self
-                .represented_battle_pet_learned_new_pet_criteria_like_cpp
-                .contains(&species)
-            {
-                self.represented_battle_pet_learned_new_pet_criteria_like_cpp
-                    .push(species);
-            }
-        }
-    }
-
     /// Finish C++ `Player::DestroyItem` after the Login DB pet/receipt commit.
     ///
     /// The receipt makes this second database phase retryable: a reconnect may
@@ -31417,530 +29484,6 @@ impl WorldSession {
         .await
     }
 
-    /// C++ `BattlePetMgr::SendError`.
-    pub(crate) fn battle_pet_send_error_like_cpp(
-        &mut self,
-        error: wow_packet::packets::misc::BattlePetErrorCodeLikeCpp,
-        creature_id: u32,
-    ) {
-        self.send_packet(&wow_packet::packets::misc::BattlePetError::new(
-            error,
-            i32::try_from(creature_id).unwrap_or(i32::MAX),
-        ));
-    }
-
-    /// C++ `BattlePetMgr::HealBattlePetsPct`.
-    ///
-    /// Fidelity note: legacy C++ does not skip removed pets and would rewrite a
-    /// damaged `BATTLE_PET_REMOVED` row to `BATTLE_PET_CHANGED`. Rust keeps the
-    /// represented removed row immutable here; if we decide to patch that legacy
-    /// bug upstream, this is the intended shared behavior.
-    #[cfg(test)]
-    pub(crate) fn battle_pet_heal_battle_pets_pct_like_cpp(&mut self, pct: u8) -> usize {
-        let mut updated = Vec::new();
-
-        for (pet_guid, pet) in &mut self.represented_battle_pets_like_cpp {
-            if pet.save_info == RepresentedBattlePetSaveInfoLikeCpp::Removed {
-                continue;
-            }
-
-            if pet.health == pet.max_health {
-                continue;
-            }
-
-            let heal = (pet.max_health as f32 * f32::from(pct) / 100.0f32) as u32;
-            pet.health = pet.health.saturating_add(heal).min(pet.max_health);
-            if pet.save_info != RepresentedBattlePetSaveInfoLikeCpp::New {
-                pet.save_info = RepresentedBattlePetSaveInfoLikeCpp::Changed;
-            }
-            updated.push(*pet_guid);
-        }
-
-        self.send_battle_pet_updates_like_cpp(&updated, false)
-    }
-
-    fn apply_battle_pet_calculated_stats_like_cpp(
-        pet: &mut RepresentedBattlePetDataLikeCpp,
-        calculated_stats: Option<RepresentedBattlePetCalculatedStatsLikeCpp>,
-    ) {
-        if let Some(calculated_stats) = calculated_stats {
-            pet.max_health = calculated_stats.max_health;
-            pet.power = calculated_stats.power;
-            pet.speed = calculated_stats.speed;
-        }
-        pet.health = pet.max_health;
-    }
-
-    /// C++ `BattlePetMgr::ChangeBattlePetQuality`. `BattlePet::CalculateStats`
-    /// may return early when breed-state DB2 rows are missing; that does not
-    /// abort the quality change.
-    #[cfg(test)]
-    pub(crate) fn battle_pet_change_battle_pet_quality_represented_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-        quality: u8,
-    ) -> RepresentedBattlePetQualityOutcomeLikeCpp {
-        if !self.has_represented_battle_pet_journal_lock_like_cpp() {
-            return RepresentedBattlePetQualityOutcomeLikeCpp::NoJournalLock;
-        }
-
-        let Some(pet) = self.represented_battle_pets_like_cpp.get(&pet_guid) else {
-            return RepresentedBattlePetQualityOutcomeLikeCpp::UnknownPet;
-        };
-
-        if quality > BATTLE_PET_BREED_QUALITY_RARE_LIKE_CPP {
-            return RepresentedBattlePetQualityOutcomeLikeCpp::QualityAboveRare;
-        }
-
-        if self.battle_pet_species_has_flag_like_cpp(
-            pet.species,
-            wow_data::BATTLE_PET_SPECIES_FLAG_CANT_BATTLE_LIKE_CPP,
-        ) {
-            return RepresentedBattlePetQualityOutcomeLikeCpp::CantBattle;
-        }
-
-        if pet.quality >= quality {
-            return RepresentedBattlePetQualityOutcomeLikeCpp::NotUpgrade;
-        }
-
-        let breed = pet.breed;
-        let species = pet.species;
-        let level = pet.level;
-        let calculated_stats =
-            self.battle_pet_calculate_stats_like_cpp(breed, species, quality, level);
-
-        let pet = self
-            .represented_battle_pets_like_cpp
-            .get_mut(&pet_guid)
-            .expect("pet was checked before stats calculation");
-        pet.quality = quality;
-        Self::apply_battle_pet_calculated_stats_like_cpp(pet, calculated_stats);
-
-        if pet.save_info != RepresentedBattlePetSaveInfoLikeCpp::New {
-            pet.save_info = RepresentedBattlePetSaveInfoLikeCpp::Changed;
-        }
-
-        self.send_battle_pet_updates_like_cpp(&[pet_guid], false);
-        RepresentedBattlePetQualityOutcomeLikeCpp::Changed
-    }
-
-    pub(crate) async fn battle_pet_change_quality_durable_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-        quality: u8,
-    ) -> RepresentedBattlePetQualityOutcomeLikeCpp {
-        let Some(attachment) = &self.battle_pet_account_attachment_like_cpp else {
-            #[cfg(test)]
-            return self
-                .battle_pet_change_battle_pet_quality_represented_like_cpp(pet_guid, quality);
-            #[cfg(not(test))]
-            return RepresentedBattlePetQualityOutcomeLikeCpp::NoJournalLock;
-        };
-        if !attachment.has_lease_like_cpp() {
-            return RepresentedBattlePetQualityOutcomeLikeCpp::NoJournalLock;
-        }
-        let owner = Arc::clone(attachment.owner_like_cpp());
-        let lease = attachment.lease_id_like_cpp();
-        let Some(pet) = owner.pet_snapshot_like_cpp(pet_guid) else {
-            return RepresentedBattlePetQualityOutcomeLikeCpp::UnknownPet;
-        };
-        if quality > BATTLE_PET_BREED_QUALITY_RARE_LIKE_CPP {
-            return RepresentedBattlePetQualityOutcomeLikeCpp::QualityAboveRare;
-        }
-        if self.battle_pet_species_has_flag_like_cpp(
-            pet.species,
-            wow_data::BATTLE_PET_SPECIES_FLAG_CANT_BATTLE_LIKE_CPP,
-        ) {
-            return RepresentedBattlePetQualityOutcomeLikeCpp::CantBattle;
-        }
-        if pet.quality >= quality {
-            return RepresentedBattlePetQualityOutcomeLikeCpp::NotUpgrade;
-        }
-        let calculated =
-            self.battle_pet_calculate_stats_like_cpp(pet.breed, pet.species, quality, pet.level);
-        match owner
-            .try_mutate_pet_like_cpp(lease, pet_guid, move |pet| {
-                pet.quality = quality;
-                Self::apply_battle_pet_calculated_stats_like_cpp(pet, calculated);
-            })
-            .await
-        {
-            Ok(((), packet)) => {
-                self.send_packet(&wow_packet::packets::misc::BattlePetUpdates {
-                    pets: vec![packet],
-                    pet_added: false,
-                });
-                RepresentedBattlePetQualityOutcomeLikeCpp::Changed
-            }
-            Err(
-                BattlePetMutationFailureLikeCpp::MissingAuthority
-                | BattlePetMutationFailureLikeCpp::JournalLocked,
-            ) => RepresentedBattlePetQualityOutcomeLikeCpp::NoJournalLock,
-            Err(BattlePetMutationFailureLikeCpp::UnknownPet) => {
-                RepresentedBattlePetQualityOutcomeLikeCpp::UnknownPet
-            }
-            Err(error) => {
-                self.log_battle_pet_mutation_failure_like_cpp("change quality", pet_guid, &error);
-                RepresentedBattlePetQualityOutcomeLikeCpp::UnknownPet
-            }
-        }
-    }
-
-    /// C++ `BattlePetMgr::GrantBattlePetLevel`. `BattlePet::CalculateStats`
-    /// may return early when breed-state DB2 rows are missing; that does not
-    /// abort the level grant.
-    #[cfg(test)]
-    pub(crate) fn battle_pet_grant_battle_pet_level_represented_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-        granted_levels: u16,
-    ) -> RepresentedBattlePetGrantLevelOutcomeLikeCpp {
-        if !self.has_represented_battle_pet_journal_lock_like_cpp() {
-            return RepresentedBattlePetGrantLevelOutcomeLikeCpp::NoJournalLock;
-        }
-
-        let Some(pet) = self.represented_battle_pets_like_cpp.get(&pet_guid) else {
-            return RepresentedBattlePetGrantLevelOutcomeLikeCpp::UnknownPet;
-        };
-
-        if self.battle_pet_species_has_flag_like_cpp(
-            pet.species,
-            wow_data::BATTLE_PET_SPECIES_FLAG_CANT_BATTLE_LIKE_CPP,
-        ) {
-            return RepresentedBattlePetGrantLevelOutcomeLikeCpp::CantBattle;
-        }
-
-        let mut level = pet.level;
-        if level >= MAX_BATTLE_PET_LEVEL_LIKE_CPP {
-            return RepresentedBattlePetGrantLevelOutcomeLikeCpp::AlreadyMaxLevel;
-        }
-
-        if granted_levels == 0 {
-            return RepresentedBattlePetGrantLevelOutcomeLikeCpp::NoGrantedLevels;
-        }
-
-        let species = pet.species;
-        let breed = pet.breed;
-        let quality = pet.quality;
-        let mut remaining_levels = granted_levels;
-        while remaining_levels > 0 && level < MAX_BATTLE_PET_LEVEL_LIKE_CPP {
-            level += 1;
-            remaining_levels -= 1;
-            #[cfg(test)]
-            self.represented_battle_pet_level_criteria_like_cpp
-                .push(RepresentedBattlePetLevelCriteriaLikeCpp { species, level });
-        }
-
-        let calculated_stats =
-            self.battle_pet_calculate_stats_like_cpp(breed, species, quality, level);
-
-        let pet = self
-            .represented_battle_pets_like_cpp
-            .get_mut(&pet_guid)
-            .expect("pet was checked before stats calculation");
-        pet.level = level;
-        if level >= MAX_BATTLE_PET_LEVEL_LIKE_CPP {
-            pet.exp = 0;
-        }
-        Self::apply_battle_pet_calculated_stats_like_cpp(pet, calculated_stats);
-
-        if pet.save_info != RepresentedBattlePetSaveInfoLikeCpp::New {
-            pet.save_info = RepresentedBattlePetSaveInfoLikeCpp::Changed;
-        }
-
-        self.send_battle_pet_updates_like_cpp(&[pet_guid], false);
-        RepresentedBattlePetGrantLevelOutcomeLikeCpp::Changed
-    }
-
-    pub(crate) async fn battle_pet_grant_level_durable_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-        granted_levels: u16,
-    ) -> RepresentedBattlePetGrantLevelOutcomeLikeCpp {
-        let Some(attachment) = &self.battle_pet_account_attachment_like_cpp else {
-            #[cfg(test)]
-            return self
-                .battle_pet_grant_battle_pet_level_represented_like_cpp(pet_guid, granted_levels);
-            #[cfg(not(test))]
-            return RepresentedBattlePetGrantLevelOutcomeLikeCpp::NoJournalLock;
-        };
-        if !attachment.has_lease_like_cpp() {
-            return RepresentedBattlePetGrantLevelOutcomeLikeCpp::NoJournalLock;
-        }
-        let owner = Arc::clone(attachment.owner_like_cpp());
-        let lease = attachment.lease_id_like_cpp();
-        let Some(pet) = owner.pet_snapshot_like_cpp(pet_guid) else {
-            return RepresentedBattlePetGrantLevelOutcomeLikeCpp::UnknownPet;
-        };
-        if self.battle_pet_species_has_flag_like_cpp(
-            pet.species,
-            wow_data::BATTLE_PET_SPECIES_FLAG_CANT_BATTLE_LIKE_CPP,
-        ) {
-            return RepresentedBattlePetGrantLevelOutcomeLikeCpp::CantBattle;
-        }
-        if pet.level >= MAX_BATTLE_PET_LEVEL_LIKE_CPP {
-            return RepresentedBattlePetGrantLevelOutcomeLikeCpp::AlreadyMaxLevel;
-        }
-        if granted_levels == 0 {
-            return RepresentedBattlePetGrantLevelOutcomeLikeCpp::NoGrantedLevels;
-        }
-        let level = pet
-            .level
-            .saturating_add(granted_levels)
-            .min(MAX_BATTLE_PET_LEVEL_LIKE_CPP);
-        #[cfg(test)]
-        let criteria: Vec<_> = ((pet.level + 1)..=level)
-            .map(|level| RepresentedBattlePetLevelCriteriaLikeCpp {
-                species: pet.species,
-                level,
-            })
-            .collect();
-        let calculated =
-            self.battle_pet_calculate_stats_like_cpp(pet.breed, pet.species, pet.quality, level);
-        match owner
-            .try_mutate_pet_like_cpp(lease, pet_guid, move |pet| {
-                pet.level = level;
-                if level >= MAX_BATTLE_PET_LEVEL_LIKE_CPP {
-                    pet.exp = 0;
-                }
-                Self::apply_battle_pet_calculated_stats_like_cpp(pet, calculated);
-            })
-            .await
-        {
-            Ok(((), packet)) => {
-                #[cfg(test)]
-                self.represented_battle_pet_level_criteria_like_cpp
-                    .extend(criteria);
-                self.send_packet(&wow_packet::packets::misc::BattlePetUpdates {
-                    pets: vec![packet],
-                    pet_added: false,
-                });
-                RepresentedBattlePetGrantLevelOutcomeLikeCpp::Changed
-            }
-            Err(
-                BattlePetMutationFailureLikeCpp::MissingAuthority
-                | BattlePetMutationFailureLikeCpp::JournalLocked,
-            ) => RepresentedBattlePetGrantLevelOutcomeLikeCpp::NoJournalLock,
-            Err(BattlePetMutationFailureLikeCpp::UnknownPet) => {
-                RepresentedBattlePetGrantLevelOutcomeLikeCpp::UnknownPet
-            }
-            Err(error) => {
-                self.log_battle_pet_mutation_failure_like_cpp("grant level", pet_guid, &error);
-                RepresentedBattlePetGrantLevelOutcomeLikeCpp::UnknownPet
-            }
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_represented_battle_pet_xp_per_level_like_cpp(
-        &mut self,
-        level: u16,
-        xp_per_level: u16,
-    ) {
-        self.represented_battle_pet_xp_per_level_like_cpp
-            .insert(level, xp_per_level);
-    }
-
-    /// C++ `BattlePetMgr::GrantBattlePetExperience`, represented after
-    /// external aura multiplier resolution.
-    #[cfg(test)]
-    pub(crate) fn battle_pet_grant_battle_pet_experience_represented_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-        xp: u16,
-        xp_source: RepresentedBattlePetXpSourceLikeCpp,
-        pet_battle_xp_multiplier: f32,
-    ) -> RepresentedBattlePetGrantExperienceOutcomeLikeCpp {
-        if !self.has_represented_battle_pet_journal_lock_like_cpp() {
-            return RepresentedBattlePetGrantExperienceOutcomeLikeCpp::NoJournalLock;
-        }
-
-        let Some(pet) = self.represented_battle_pets_like_cpp.get(&pet_guid) else {
-            return RepresentedBattlePetGrantExperienceOutcomeLikeCpp::UnknownPet;
-        };
-
-        if xp == 0 || xp_source == RepresentedBattlePetXpSourceLikeCpp::Invalid {
-            return RepresentedBattlePetGrantExperienceOutcomeLikeCpp::InvalidXpOrSource;
-        }
-
-        if self.battle_pet_species_has_flag_like_cpp(
-            pet.species,
-            wow_data::BATTLE_PET_SPECIES_FLAG_CANT_BATTLE_LIKE_CPP,
-        ) {
-            return RepresentedBattlePetGrantExperienceOutcomeLikeCpp::CantBattle;
-        }
-
-        let mut level = pet.level;
-        if level >= MAX_BATTLE_PET_LEVEL_LIKE_CPP {
-            return RepresentedBattlePetGrantExperienceOutcomeLikeCpp::AlreadyMaxLevel;
-        }
-
-        let Some(mut next_level_xp) = self.battle_pet_xp_per_level_like_cpp(level) else {
-            return RepresentedBattlePetGrantExperienceOutcomeLikeCpp::MissingXpRow;
-        };
-
-        let species = pet.species;
-        let breed = pet.breed;
-        let quality = pet.quality;
-        let mut total_xp = if xp_source == RepresentedBattlePetXpSourceLikeCpp::PetBattle {
-            (f32::from(xp) * pet_battle_xp_multiplier) as u16
-        } else {
-            xp
-        };
-        total_xp = total_xp.saturating_add(pet.exp);
-
-        while total_xp >= next_level_xp && level < MAX_BATTLE_PET_LEVEL_LIKE_CPP {
-            total_xp = total_xp.saturating_sub(next_level_xp);
-            level += 1;
-
-            let Some(row_xp) = self.battle_pet_xp_per_level_like_cpp(level) else {
-                return RepresentedBattlePetGrantExperienceOutcomeLikeCpp::MissingXpRow;
-            };
-            next_level_xp = row_xp;
-
-            #[cfg(test)]
-            {
-                let criteria = RepresentedBattlePetLevelCriteriaLikeCpp { species, level };
-                self.represented_battle_pet_level_criteria_like_cpp
-                    .push(criteria);
-                if xp_source == RepresentedBattlePetXpSourceLikeCpp::PetBattle {
-                    self.represented_battle_pet_active_level_criteria_like_cpp
-                        .push(criteria);
-                }
-            }
-        }
-
-        let calculated_stats =
-            self.battle_pet_calculate_stats_like_cpp(breed, species, quality, level);
-
-        let pet = self
-            .represented_battle_pets_like_cpp
-            .get_mut(&pet_guid)
-            .expect("pet was checked before XP calculation");
-        pet.level = level;
-        pet.exp = if level < MAX_BATTLE_PET_LEVEL_LIKE_CPP {
-            total_xp
-        } else {
-            0
-        };
-        Self::apply_battle_pet_calculated_stats_like_cpp(pet, calculated_stats);
-
-        if pet.save_info != RepresentedBattlePetSaveInfoLikeCpp::New {
-            pet.save_info = RepresentedBattlePetSaveInfoLikeCpp::Changed;
-        }
-
-        self.send_battle_pet_updates_like_cpp(&[pet_guid], false);
-        RepresentedBattlePetGrantExperienceOutcomeLikeCpp::Changed
-    }
-
-    pub(crate) async fn battle_pet_grant_experience_durable_like_cpp(
-        &mut self,
-        pet_guid: ObjectGuid,
-        xp: u16,
-        xp_source: RepresentedBattlePetXpSourceLikeCpp,
-        pet_battle_xp_multiplier: f32,
-    ) -> RepresentedBattlePetGrantExperienceOutcomeLikeCpp {
-        let Some(attachment) = &self.battle_pet_account_attachment_like_cpp else {
-            #[cfg(test)]
-            return self.battle_pet_grant_battle_pet_experience_represented_like_cpp(
-                pet_guid,
-                xp,
-                xp_source,
-                pet_battle_xp_multiplier,
-            );
-            #[cfg(not(test))]
-            return RepresentedBattlePetGrantExperienceOutcomeLikeCpp::NoJournalLock;
-        };
-        if !attachment.has_lease_like_cpp() {
-            return RepresentedBattlePetGrantExperienceOutcomeLikeCpp::NoJournalLock;
-        }
-        let owner = Arc::clone(attachment.owner_like_cpp());
-        let lease = attachment.lease_id_like_cpp();
-        let Some(pet) = owner.pet_snapshot_like_cpp(pet_guid) else {
-            return RepresentedBattlePetGrantExperienceOutcomeLikeCpp::UnknownPet;
-        };
-        if xp == 0 || xp_source == RepresentedBattlePetXpSourceLikeCpp::Invalid {
-            return RepresentedBattlePetGrantExperienceOutcomeLikeCpp::InvalidXpOrSource;
-        }
-        if self.battle_pet_species_has_flag_like_cpp(
-            pet.species,
-            wow_data::BATTLE_PET_SPECIES_FLAG_CANT_BATTLE_LIKE_CPP,
-        ) {
-            return RepresentedBattlePetGrantExperienceOutcomeLikeCpp::CantBattle;
-        }
-        if pet.level >= MAX_BATTLE_PET_LEVEL_LIKE_CPP {
-            return RepresentedBattlePetGrantExperienceOutcomeLikeCpp::AlreadyMaxLevel;
-        }
-        let Some(mut next_level_xp) = self.battle_pet_xp_per_level_like_cpp(pet.level) else {
-            return RepresentedBattlePetGrantExperienceOutcomeLikeCpp::MissingXpRow;
-        };
-        let mut level = pet.level;
-        let mut total_xp = if xp_source == RepresentedBattlePetXpSourceLikeCpp::PetBattle {
-            (f32::from(xp) * pet_battle_xp_multiplier) as u16
-        } else {
-            xp
-        };
-        total_xp = total_xp.saturating_add(pet.exp);
-        #[cfg(test)]
-        let mut criteria = Vec::new();
-        while total_xp >= next_level_xp && level < MAX_BATTLE_PET_LEVEL_LIKE_CPP {
-            total_xp = total_xp.saturating_sub(next_level_xp);
-            level += 1;
-            let Some(row_xp) = self.battle_pet_xp_per_level_like_cpp(level) else {
-                return RepresentedBattlePetGrantExperienceOutcomeLikeCpp::MissingXpRow;
-            };
-            next_level_xp = row_xp;
-            #[cfg(test)]
-            criteria.push(RepresentedBattlePetLevelCriteriaLikeCpp {
-                species: pet.species,
-                level,
-            });
-        }
-        let calculated =
-            self.battle_pet_calculate_stats_like_cpp(pet.breed, pet.species, pet.quality, level);
-        let persisted_exp = if level < MAX_BATTLE_PET_LEVEL_LIKE_CPP {
-            total_xp
-        } else {
-            0
-        };
-        match owner
-            .try_mutate_pet_like_cpp(lease, pet_guid, move |pet| {
-                pet.level = level;
-                pet.exp = persisted_exp;
-                Self::apply_battle_pet_calculated_stats_like_cpp(pet, calculated);
-            })
-            .await
-        {
-            Ok(((), packet)) => {
-                #[cfg(test)]
-                {
-                    self.represented_battle_pet_level_criteria_like_cpp
-                        .extend(criteria.iter().copied());
-                    if xp_source == RepresentedBattlePetXpSourceLikeCpp::PetBattle {
-                        self.represented_battle_pet_active_level_criteria_like_cpp
-                            .extend(criteria);
-                    }
-                }
-                self.send_packet(&wow_packet::packets::misc::BattlePetUpdates {
-                    pets: vec![packet],
-                    pet_added: false,
-                });
-                RepresentedBattlePetGrantExperienceOutcomeLikeCpp::Changed
-            }
-            Err(
-                BattlePetMutationFailureLikeCpp::MissingAuthority
-                | BattlePetMutationFailureLikeCpp::JournalLocked,
-            ) => RepresentedBattlePetGrantExperienceOutcomeLikeCpp::NoJournalLock,
-            Err(BattlePetMutationFailureLikeCpp::UnknownPet) => {
-                RepresentedBattlePetGrantExperienceOutcomeLikeCpp::UnknownPet
-            }
-            Err(error) => {
-                self.log_battle_pet_mutation_failure_like_cpp("grant experience", pet_guid, &error);
-                RepresentedBattlePetGrantExperienceOutcomeLikeCpp::UnknownPet
-            }
-        }
-    }
-
     #[cfg(test)]
     pub(crate) fn battle_pet_grant_battle_pet_experience_with_owner_auras_like_cpp(
         &mut self,
@@ -31959,263 +29502,6 @@ impl WorldSession {
         self.battle_pet_grant_battle_pet_experience_represented_like_cpp(
             pet_guid, xp, xp_source, multiplier,
         )
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_battle_pet_level_criteria_like_cpp(
-        &self,
-    ) -> &[RepresentedBattlePetLevelCriteriaLikeCpp] {
-        &self.represented_battle_pet_level_criteria_like_cpp
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_battle_pet_active_level_criteria_like_cpp(
-        &self,
-    ) -> &[RepresentedBattlePetLevelCriteriaLikeCpp] {
-        &self.represented_battle_pet_active_level_criteria_like_cpp
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_battle_pet_unique_owned_criteria_like_cpp(&self) -> u32 {
-        self.represented_battle_pet_unique_owned_criteria_like_cpp
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_battle_pet_learned_new_pet_criteria_like_cpp(&self) -> &[u32] {
-        &self.represented_battle_pet_learned_new_pet_criteria_like_cpp
-    }
-
-    /// Publish one complete fallback projection of
-    /// `battle_pet_slots(id, battlePetGuid, locked)`. The canonical account
-    /// attachment owns this authority in production; this seam exists for
-    /// isolated represented tests and rejects missing, duplicate, or
-    /// out-of-range slot rows.
-    #[cfg(test)]
-    pub(crate) fn complete_represented_battle_pet_slot_authority_load_like_cpp(
-        &mut self,
-        rows: impl IntoIterator<Item = (u8, Option<ObjectGuid>, bool)>,
-    ) -> bool {
-        self.invalidate_canonical_player_spell_hit_aura_authority_like_cpp();
-        self.represented_battle_pet_slots_authority_complete_like_cpp = false;
-
-        let mut slots =
-            std::array::from_fn(|index| RepresentedBattlePetSlotLikeCpp::locked_empty(index as u8));
-        let mut seen = [false; BATTLE_PET_SLOT_COUNT_LIKE_CPP];
-        for (index, pet_guid, locked) in rows {
-            let slot_index = usize::from(index);
-            let Some(slot) = slots.get_mut(slot_index) else {
-                return false;
-            };
-            if seen[slot_index] {
-                return false;
-            }
-            seen[slot_index] = true;
-            *slot = RepresentedBattlePetSlotLikeCpp {
-                pet_guid,
-                collar_id: 0,
-                index,
-                locked,
-            };
-        }
-        if !seen.into_iter().all(|present| present) {
-            return false;
-        }
-
-        self.represented_battle_pet_slots_like_cpp = slots;
-        self.represented_battle_pet_slots_authority_complete_like_cpp = true;
-        true
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_battle_pet_slot_like_cpp(&self, slot: u8) -> Option<ObjectGuid> {
-        self.represented_battle_pet_slots_like_cpp
-            .get(slot as usize)
-            .and_then(|slot| slot.pet_guid)
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_battle_pet_slot_locked_like_cpp(&self, slot: u8) -> Option<bool> {
-        self.represented_battle_pet_slots_like_cpp
-            .get(slot as usize)
-            .map(|slot| slot.locked)
-    }
-
-    /// C++ `BattlePetMgr::SendJournal` packet body builder.
-    pub(crate) fn represented_battle_pet_journal_like_cpp(
-        &self,
-    ) -> Option<wow_packet::packets::misc::BattlePetJournal> {
-        if let Some(attachment) = &self.battle_pet_account_attachment_like_cpp {
-            return Some(
-                attachment
-                    .owner_like_cpp()
-                    .journal_like_cpp(attachment.lease_id_like_cpp(), self.player_guid()),
-            );
-        }
-        #[cfg(not(test))]
-        return None;
-        #[cfg(test)]
-        {
-            let player_guid = self.player_guid();
-            let mut journal = wow_packet::packets::misc::BattlePetJournal {
-                trap: 0,
-                has_journal_lock: self.has_represented_battle_pet_journal_lock_like_cpp(),
-                slots: Vec::with_capacity(BATTLE_PET_SLOT_COUNT_LIKE_CPP),
-                pets: Vec::new(),
-            };
-
-            for (pet_guid, pet) in &self.represented_battle_pets_like_cpp {
-                if pet.save_info == RepresentedBattlePetSaveInfoLikeCpp::Removed {
-                    continue;
-                }
-
-                if pet
-                    .owner_info
-                    .is_some_and(|owner_info| Some(owner_info.guid) != player_guid)
-                {
-                    continue;
-                }
-
-                journal.pets.push(pet.packet_info_like_cpp(*pet_guid));
-            }
-
-            for slot in &self.represented_battle_pet_slots_like_cpp {
-                let mut packet_slot = slot.packet_slot_like_cpp();
-                if packet_slot.pet_guid
-                    != wow_packet::packets::misc::empty_battle_pet_guid_like_cpp()
-                    && !self
-                        .represented_battle_pets_like_cpp
-                        .contains_key(&packet_slot.pet_guid)
-                {
-                    packet_slot.pet_guid =
-                        wow_packet::packets::misc::empty_battle_pet_guid_like_cpp();
-                }
-                journal.slots.push(packet_slot);
-            }
-
-            Some(journal)
-        }
-    }
-
-    /// C++ `WorldSession::HandleBattlePetSummon` represented toggle.
-    ///
-    /// `BattlePetMgr::SummonPet` silently ignores unknown pets before casting
-    /// the summon spell; `DismissPet` clears the active summoned companion.
-    pub(crate) fn battle_pet_summon_toggle_like_cpp(&mut self, pet_guid: ObjectGuid) -> bool {
-        if self.represented_summoned_battle_pet_guid_like_cpp() == Some(pet_guid) {
-            return self.set_represented_summoned_battle_pet_guid_like_cpp(None);
-        }
-
-        if self.represented_battle_pet_like_cpp(pet_guid).is_none() {
-            return false;
-        }
-
-        self.set_represented_summoned_battle_pet_guid_like_cpp(Some(pet_guid))
-    }
-
-    pub(crate) fn represented_summoned_battle_pet_guid_like_cpp(&self) -> Option<ObjectGuid> {
-        if let Some(guid) =
-            self.with_owned_player_like_cpp(Player::summoned_battle_pet_guid_like_cpp)
-        {
-            return guid;
-        }
-        #[cfg(test)]
-        {
-            self.represented_summoned_battle_pet_guid_like_cpp
-        }
-        #[cfg(not(test))]
-        {
-            None
-        }
-    }
-
-    fn set_represented_summoned_battle_pet_guid_like_cpp(
-        &mut self,
-        pet_guid: Option<ObjectGuid>,
-    ) -> bool {
-        if self
-            .with_owned_player_mut_like_cpp(|player| {
-                player.set_summoned_battle_pet_guid_like_cpp(
-                    pet_guid.unwrap_or(wow_core::ObjectGuid::EMPTY),
-                );
-            })
-            .is_some()
-        {
-            return true;
-        }
-        #[cfg(test)]
-        {
-            self.represented_summoned_battle_pet_guid_like_cpp = pet_guid;
-            true
-        }
-        #[cfg(not(test))]
-        {
-            false
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_battle_pet_cage_items_like_cpp(
-        &self,
-    ) -> &[RepresentedBattlePetCageItemLikeCpp] {
-        &self.represented_battle_pet_cage_items_like_cpp
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_represented_battle_pet_query_companion_like_cpp(
-        &mut self,
-        unit_guid: ObjectGuid,
-        companion: RepresentedBattlePetQueryCompanionLikeCpp,
-    ) {
-        self.represented_battle_pet_query_companions_like_cpp
-            .insert(unit_guid, companion);
-    }
-
-    pub(crate) fn represented_battle_pet_query_companion_like_cpp(
-        &self,
-        unit_guid: ObjectGuid,
-    ) -> Option<RepresentedBattlePetQueryCompanionLikeCpp> {
-        if let Some(manager) = self.canonical_map_manager.as_ref()
-            && let Ok(manager) = manager.lock()
-        {
-            let mut snapshot = None;
-            manager.do_for_all_maps(|managed| {
-                if snapshot.is_some() {
-                    return;
-                }
-                let Some(companion) = managed.map().with_creature_like_cpp(unit_guid, |creature| {
-                    let unit = creature.unit();
-                    RepresentedBattlePetQueryCompanionLikeCpp {
-                        creature_id: i32::try_from(creature.entry()).unwrap_or(i32::MAX),
-                        name_timestamp: i64::from(
-                            unit.battle_pet_companion_name_timestamp_like_cpp(),
-                        ),
-                        is_summon: creature.is_summon_like_cpp(),
-                        owner_is_player: unit
-                            .subsystems()
-                            .control
-                            .owner_guid
-                            .is_some_and(|guid| guid.is_player()),
-                        battle_pet_companion_guid: unit.battle_pet_companion_guid_like_cpp(),
-                    }
-                }) else {
-                    return;
-                };
-                snapshot = Some(companion);
-            });
-            if snapshot.is_some() {
-                return snapshot;
-            }
-        }
-        #[cfg(test)]
-        {
-            self.represented_battle_pet_query_companions_like_cpp
-                .get(&unit_guid)
-                .copied()
-        }
-        #[cfg(not(test))]
-        {
-            None
-        }
     }
 
     pub(crate) fn set_represented_critter_guid_like_cpp(&mut self, guid: Option<ObjectGuid>) {
@@ -32280,46 +29566,6 @@ impl WorldSession {
     #[cfg(test)]
     pub(crate) fn represented_dismissed_critter_guids_like_cpp(&self) -> &[ObjectGuid] {
         &self.represented_dismissed_critter_guids_like_cpp
-    }
-
-    /// C++ `BattlePetMgr::UpdateBattlePetData`, represented at the gate level.
-    pub(crate) fn battle_pet_update_notify_like_cpp(&mut self, pet_guid: ObjectGuid) -> bool {
-        let Some(pet) = self.represented_battle_pet_like_cpp(pet_guid) else {
-            return false;
-        };
-
-        if self.represented_summoned_battle_pet_guid_like_cpp() != Some(pet_guid) {
-            return false;
-        }
-
-        let _ = self.mutate_canonical_player_like_cpp(|player| {
-            player.set_battle_pet_data_like_cpp(pet_guid, pet.quality, pet.level);
-        });
-        #[cfg(test)]
-        self.represented_battle_pet_data_updates_like_cpp
-            .push(pet_guid);
-        true
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_battle_pet_data_updates_like_cpp(&self) -> &[ObjectGuid] {
-        &self.represented_battle_pet_data_updates_like_cpp
-    }
-
-    pub(crate) fn represented_battle_pet_like_cpp(
-        &self,
-        pet_guid: ObjectGuid,
-    ) -> Option<RepresentedBattlePetDataLikeCpp> {
-        if let Some(attachment) = &self.battle_pet_account_attachment_like_cpp {
-            return attachment.owner_like_cpp().pet_snapshot_like_cpp(pet_guid);
-        }
-        #[cfg(test)]
-        return self
-            .represented_battle_pets_like_cpp
-            .get(&pet_guid)
-            .cloned();
-        #[cfg(not(test))]
-        None
     }
 
     pub(crate) fn represented_player_reject_battleground_object_vehicle_like_cpp(
@@ -33729,78 +30975,6 @@ impl WorldSession {
         self.delayed_operations_processed_like_cpp
     }
 
-    fn represented_pet_position_like_cpp(&self, pet_guid: ObjectGuid) -> Option<Position> {
-        let map_id = u32::from(self.player_map_id_like_cpp());
-        let instance_id = self
-            .current_canonical_player_map_key_like_cpp()
-            .map(|key| key.instance_id)
-            .unwrap_or(0);
-        let manager = self.canonical_map_manager.as_ref()?.lock().ok()?;
-        let managed = manager.find_map(map_id, instance_id)?;
-        managed.map().with_world_object_by_kinds_like_cpp(
-            pet_guid,
-            &[AccessorObjectKind::Pet],
-            |object| object.position(),
-        )
-    }
-
-    fn send_represented_pet_spline_speed_like_cpp(
-        &self,
-        pet_guid: ObjectGuid,
-        move_type: UnitMoveTypeLikeCpp,
-        rate: f32,
-    ) {
-        let Some(opcode) = Self::creature_movement_spline_speed_opcode_like_cpp(move_type) else {
-            return;
-        };
-        let packet_bytes = wow_packet::packets::movement::MoveSplineSetSpeed {
-            opcode,
-            mover_guid: pet_guid,
-            speed: PLAYER_BASE_MOVE_SPEED_LIKE_CPP[move_type.index()] * rate,
-        }
-        .to_bytes();
-        let map_id = self.player_map_id_like_cpp();
-        let instance_id = self
-            .current_canonical_player_map_key_like_cpp()
-            .map(|key| key.instance_id)
-            .unwrap_or(0);
-
-        if self.client_visible_guids_like_cpp.contains(&pet_guid)
-            && self.send_tx().send(packet_bytes.clone()).is_err()
-        {
-            warn!("Send channel closed for account {}", self.account_id);
-        }
-
-        let (Some(player_guid), Some(registry)) = (self.player_guid(), self.player_registry())
-        else {
-            return;
-        };
-        let Some(source_position) = self
-            .represented_pet_position_like_cpp(pet_guid)
-            .or_else(|| self.player_position_like_cpp())
-        else {
-            return;
-        };
-        for registration in registry.movement_recipients_within_range(
-            player_guid,
-            map_id,
-            instance_id,
-            source_position,
-            crate::map_manager::VISIBILITY_RADIUS,
-        ) {
-            let _ = registry.try_send_current_command(
-                registration,
-                SessionCommand::SendIfVisibleLikeCpp(SendIfVisibleLikeCppCommand {
-                    queued_at: Instant::now(),
-                    source_guid: pet_guid,
-                    map_id,
-                    instance_id,
-                    packet_bytes: packet_bytes.clone(),
-                }),
-            );
-        }
-    }
-
     fn set_represented_can_fly_like_cpp(&mut self, enable: bool) -> bool {
         let Some(mut movement_flags) = self.resolved_player_movement_flags_like_cpp() else {
             return false;
@@ -33871,60 +31045,6 @@ impl WorldSession {
         true
     }
 
-    fn propagate_represented_player_speed_to_pet_like_cpp(
-        &mut self,
-        move_type: UnitMoveTypeLikeCpp,
-        rate: f32,
-    ) {
-        let Some(pet_guid) = self.player_pet_guid_state_like_cpp().flatten() else {
-            return;
-        };
-        if self.resolved_in_combat_like_cpp() != Some(false) {
-            return;
-        }
-
-        let rate = rate.max(0.01);
-        let index = move_type.index();
-        let canonical_changed = self.with_canonical_pet_mut_like_cpp(pet_guid, |pet| {
-            let unit = pet.creature_mut().unit_mut();
-            if unit.speed_rate_at_like_cpp(index) == Some(rate) {
-                return false;
-            }
-            unit.set_speed_rate_at_like_cpp(index, rate)
-        });
-        let changed = match canonical_changed {
-            Some(changed) => changed,
-            None => {
-                #[cfg(test)]
-                {
-                    if self.player_handle_like_cpp.is_none() {
-                        if self.represented_pet_movement_speed_rates_like_cpp[index] == rate {
-                            return;
-                        }
-                        self.represented_pet_movement_speed_rates_like_cpp[index] = rate;
-                        true
-                    } else {
-                        false
-                    }
-                }
-                #[cfg(not(test))]
-                {
-                    false
-                }
-            }
-        };
-        if !changed {
-            return;
-        }
-        #[cfg(test)]
-        {
-            self.represented_pet_speed_propagations_like_cpp = self
-                .represented_pet_speed_propagations_like_cpp
-                .saturating_add(1);
-        }
-        self.send_represented_pet_spline_speed_like_cpp(pet_guid, move_type, rate);
-    }
-
     pub(crate) fn trace_anticheat_violation_like_cpp(
         &self,
         rule: &'static str,
@@ -33940,30 +31060,6 @@ impl WorldSession {
             severity,
             "anticheat.violation"
         );
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_pet_movement_speed_rate_like_cpp(
-        &self,
-        move_type: UnitMoveTypeLikeCpp,
-    ) -> f32 {
-        let canonical = self
-            .player_pet_guid_state_like_cpp()
-            .flatten()
-            .and_then(|pet_guid| {
-                self.with_canonical_pet_like_cpp(pet_guid, |pet| {
-                    pet.creature()
-                        .unit()
-                        .speed_rate_at_like_cpp(move_type.index())
-                })
-                .flatten()
-            });
-        canonical.unwrap_or(self.represented_pet_movement_speed_rates_like_cpp[move_type.index()])
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_pet_speed_propagations_like_cpp(&self) -> u32 {
-        self.represented_pet_speed_propagations_like_cpp
     }
 
     pub(crate) fn set_player_transport_guid_like_cpp(&mut self, guid: Option<ObjectGuid>) {
