+++ tools/architecture/physical-file-policy.json
@@ -629,9 +629,9 @@
     },
     {
       "path": "crates/wow-world/src/session/mod.rs",
-      "observed_lines": 47905,
-      "ceiling_lines": 47905,
-      "split": "Retire gameplay orchestration/catalog access by complete operations. #585 adds the task-owned finalization ledger and identity-hydration guard. PR #586 review adds 16 inseparable lines to retain the native far-transfer destination before an interruptible realm writer fence in both immediate and delayed paths. #589 moves the represented Player cast publication out to session/player_cast/, a net reduction of 267 lines; the ceiling is tightened to the validated 75840. Existing update methods become async in the same Session task; no new field, task, lock or state mirror. New failure/cancellation tests remain in bounded files. This reviewed growth is not physical closeout: the remaining C4 root split stays in #584. #597 separates the represented item and inventory responsibility into 18 bounded session/player_items submodules, a validated reduction from 75840 to 67187 lines; the ceiling is tightened to that figure. The move is method-for-method with an identical impl surface before and after, so this is physical separation, not a behaviour change. The remaining C4 root split stays in #584. #599 separates the represented world-entity responsibility - creature, gameobject, aggro and spawn - into 18 bounded session/world_entities submodules, a validated reduction from 66,808 to 60,041 lines; the ceiling is tightened to that figure and the stale 67,187 entry left by #597 is corrected to the real count. The move is method-for-method with an identical impl surface of 2,149 methods before and after. #601 separates the represented aura and spell-state responsibility into 14 bounded session/spell_state submodules, a validated reduction from 60,041 to 54,029 lines; the ceiling is tightened to that figure. Method-for-method with an identical impl surface of 1,983 methods before and after. #603 separates the represented movement and world-transfer responsibility into 7 bounded session/movement submodules, a validated reduction from 54,029 to 50,802 lines; the ceiling is tightened to that figure. Method-for-method with an identical impl surface of 1,714 methods before and after; the #585/#586 far-transfer writer fence ordering is relocated, not altered. #605 separates the represented quest responsibility into 7 bounded session/quest submodules, a validated reduction from 50,802 to 47,905 lines; the ceiling is tightened to that figure. Method-for-method with an identical impl surface of 1,577 methods before and after.",
+      "observed_lines": 45001,
+      "ceiling_lines": 45001,
+      "split": "Retire gameplay orchestration/catalog access by complete operations. #585 adds the task-owned finalization ledger and identity-hydration guard. PR #586 review adds 16 inseparable lines to retain the native far-transfer destination before an interruptible realm writer fence in both immediate and delayed paths. #589 moves the represented Player cast publication out to session/player_cast/, a net reduction of 267 lines; the ceiling is tightened to the validated 75840. Existing update methods become async in the same Session task; no new field, task, lock or state mirror. New failure/cancellation tests remain in bounded files. This reviewed growth is not physical closeout: the remaining C4 root split stays in #584. #597 separates the represented item and inventory responsibility into 18 bounded session/player_items submodules, a validated reduction from 75840 to 67187 lines; the ceiling is tightened to that figure. The move is method-for-method with an identical impl surface before and after, so this is physical separation, not a behaviour change. The remaining C4 root split stays in #584. #599 separates the represented world-entity responsibility - creature, gameobject, aggro and spawn - into 18 bounded session/world_entities submodules, a validated reduction from 66,808 to 60,041 lines; the ceiling is tightened to that figure and the stale 67,187 entry left by #597 is corrected to the real count. The move is method-for-method with an identical impl surface of 2,149 methods before and after. #601 separates the represented aura and spell-state responsibility into 14 bounded session/spell_state submodules, a validated reduction from 60,041 to 54,029 lines; the ceiling is tightened to that figure. Method-for-method with an identical impl surface of 1,983 methods before and after. #603 separates the represented movement and world-transfer responsibility into 7 bounded session/movement submodules, a validated reduction from 54,029 to 50,802 lines; the ceiling is tightened to that figure. Method-for-method with an identical impl surface of 1,714 methods before and after; the #585/#586 far-transfer writer fence ordering is relocated, not altered. #605 separates the represented quest responsibility into 7 bounded session/quest submodules, a validated reduction from 50,802 to 47,905 lines; the ceiling is tightened to that figure. Method-for-method with an identical impl surface of 1,577 methods before and after. #607 separates the represented pet and battle-pet responsibility into 7 bounded session/pets submodules, a validated reduction from 47,905 to 45,001 lines; the ceiling is tightened to that figure. Method-for-method with an identical impl surface.",
       "review_checkpoint": "584:C4"
     },
     {
